+++ nomenklatura/wikidata/client.py
@@ -78,7 +78,7 @@ class WikidataClient(object):
     def query(self, query_text: str) -> SparqlResponse:
         """Query the Wikidata SPARQL endpoint."""
         clean_text: Optional[str] = collapse_spaces(query_text)
-        if clean_text is None:
+        if clean_text is None or len(clean_text) == 0:
             raise RuntimeError("Invalid query: %r" % query_text)
         params = {"query": clean_text}
         url = build_url(self.QUERY_API, params=params)
