+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q104086571.json
@@ -1,0 +1,1 @@
+{"entities": {"Q104086571": {"pageid": 99560591, "ns": 0, "title": "Q104086571", "lastrevid": 2309068621, "modified": "2025-02-09T11:55:43Z", "type": "item", "id": "Q104086571", "labels": {"en": {"language": "en", "value": "first-order class"}}, "descriptions": {"en": {"language": "en", "value": "class containing as instances all classes whose instances are individuals and not classes"}}, "aliases": {"en": [{"language": "en", "value": "token-class"}, {"language": "en", "value": "class of individuals"}, {"language": "en", "value": "first-order type"}, {"language": "en", "value": "1stOT"}, {"language": "en", "value": "metaclass of the first order"}, {"language": "en", "value": "metaclass of the 1st order"}, {"language": "en", "value": "1st-order metaclass"}, {"language": "en", "value": "metaclass (1st order)"}, {"language": "en", "value": "first-order metaclass"}, {"language": "en", "value": "metaclass (first order)"}, {"language": "en", "value": "0-order metaclass"}, {"language": "en", "value": "Wikidata instance class"}, {"language": "en", "value": "Wikidata simple class"}, {"language": "en", "value": "class"}]}, "claims": {"P31": [{"mainsnak": {"snaktype": "value", "property": "P31", "hash": "a7e32719e3a70a25fce867c2ad70c306d3521fc3", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24017414, "id": "Q24017414"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q104086571$28b768d1-45e4-3bde-5a22-0a6bd4adb229", "rank": "normal"}], "P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "c1a4184b5aab81dd72dd84a384a7ab3522eaa365", "datavalue": {"value": {"entity-type": "item", "numeric-id": 23959932, "id": "Q23959932"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P1545": [{"snaktype": "value", "property": "P1545", "hash": "2a1ced1dca90648ea7e306acbadd74fc81a10722", "datavalue": {"value": "1", "type": "string"}, "datatype": "string"}], "P156": [{"snaktype": "value", "property": "P156", "hash": "3b1bdf9ccc394a2b181784ffc8ce79a32ec34e26", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24017414, "id": "Q24017414"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P1545", "P156"], "id": "Q104086571$1a02556c-48aa-1a98-898b-73d952f1d333", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P279", "hash": "33c0fd1431aa0fb276f87bb51b4a956d9440bb15", "datavalue": {"value": {"entity-type": "item", "numeric-id": 16889133, "id": "Q16889133"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q104086571$1818AEF8-C491-4B01-8CB2-FF71A4149427", "rank": "normal"}], "P2559": [{"mainsnak": {"snaktype": "value", "property": "P2559", "hash": "7b4555f5a016c33322e4e2e8b0b8eb703199e86a", "datavalue": {"value": {"text": "A first-order class in Wikidata, e.g., watercraft (Q1229765), bridge (Q12280), or human (Q5), is a class none of whose instances can be classes. Class that are a subclass of (P279) another class that is a first-order class are implicitly a first-order class, e.g., submarine (Q2811) is implicitly a first order class because it is a subclass of (P279) of (a subclass of) watercraft (Q1229765). There are many classes in Wikidata whose instances cannot be classes. It is only necessary to state that such a class is an instance of (P31) first-order class (Q104086571) to emphasize this point, particularly to distinguish it from a related second-order class such as to distinguish between watercraft (Q1229765) and watercraft type (Q16335899). Before stating that a class is an instance of (P31) first-order class (Q104086571) check that it is actually a first-order class, i.e., that it cannot have any instances that are classes.", "language": "en"}, "type": "monolingualtext"}, "datatype": "monolingualtext"}, "type": "statement", "id": "Q104086571$62bbe0dc-4587-f301-ba52-9328a7dabc72", "rank": "normal"}], "P6104": [{"mainsnak": {"snaktype": "value", "property": "P6104", "hash": "2bcf963484ee556d691849e71ea3ae951738fc36", "datavalue": {"value": {"entity-type": "item", "numeric-id": 60005226, "id": "Q60005226"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q104086571$259DC0F7-060C-4367-B364-5FBC40034004", "rank": "normal"}]}}}, "success": 1}
