+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q10541491.json
@@ -1,0 +1,1 @@
+{"entities": {"Q10541491": {"pageid": 11833471, "ns": 0, "title": "Q10541491", "lastrevid": 2322064247, "modified": "2025-03-09T08:45:34Z", "type": "item", "id": "Q10541491", "labels": {"en": {"language": "en", "value": "legal form"}}, "descriptions": {"en": {"language": "en", "value": "type of legal entity within certain legal system"}}, "aliases": {}, "claims": {"P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "33c0fd1431aa0fb276f87bb51b4a956d9440bb15", "datavalue": {"value": {"entity-type": "item", "numeric-id": 16889133, "id": "Q16889133"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q10541491$943367E2-6DE1-4DB7-AC5D-167ED197068E", "rank": "normal"}], "P227": [{"mainsnak": {"snaktype": "value", "property": "P227", "hash": "67ee37e7438483a710cb37ef24325145315047f3", "datavalue": {"value": "4133316-0", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q10541491$907A3F09-BBC2-4456-B9B6-924AFD986692", "rank": "normal"}], "P31": [{"mainsnak": {"snaktype": "value", "property": "P31", "hash": "a7e32719e3a70a25fce867c2ad70c306d3521fc3", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24017414, "id": "Q24017414"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q10541491$6C835297-A687-496F-8227-BB924E833DBB", "rank": "normal"}], "P2671": [{"mainsnak": {"snaktype": "value", "property": "P2671", "hash": "3af1e981f5f098bad5fa93f4b2976f568177dda4", "datavalue": {"value": "/g/121y9xhf", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q10541491$D472C367-BDDC-4E50-9267-5BD6355D8C07", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P2671", "hash": "39a18d0ce6b1784e17511c7039f1baeb6ac94feb", "datavalue": {"value": "/g/1234zw0y", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q10541491$D18EC8B4-DB79-4807-8FF4-5B027389508C", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P2671", "hash": "00291a5bea0412d87447e6b2b0accf168f210d98", "datavalue": {"value": "/g/1227csdp", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q10541491$CBE6A6BC-8CA1-4BCB-89D9-186E93AE80CF", "rank": "normal"}], "P3911": [{"mainsnak": {"snaktype": "value", "property": "P3911", "hash": "f568dca8dafaa27767c73ead2d5f14a3cf751eb6", "datavalue": {"value": "12194-1", "type": "string"}, "datatype": "external-id"}, "type": "statement", "qualifiers": {"P4390": [{"snaktype": "value", "property": "P4390", "hash": "1a4df62914ea9afca349bd5fb5d8efd5832d83fa", "datavalue": {"value": {"entity-type": "item", "numeric-id": 39893449, "id": "Q39893449"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P4390"], "id": "Q10541491$08471CA4-88E8-4336-AB1A-7AEABFB69970", "rank": "normal"}], "P2888": [{"mainsnak": {"snaktype": "value", "property": "P2888", "hash": "16ac25ae0f2093ecb3e7107816ed5e0400f1219e", "datavalue": {"value": "https://w3id.org/italia/onto/COV/LegalStatus", "type": "string"}, "datatype": "url"}, "type": "statement", "id": "Q10541491$f5db8806-4a28-be2a-2f9a-66e6f9d95d9a", "rank": "normal"}], "P1687": [{"mainsnak": {"snaktype": "value", "property": "P1687", "hash": "040ad2184f12f00c2ea486b71b9de6c7a92f2541", "datavalue": {"value": {"entity-type": "property", "numeric-id": 1454, "id": "P1454"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}, "type": "statement", "id": "Q10541491$42741fe5-45f1-5c90-3a29-38eecf6d6512", "rank": "normal"}], "P910": [{"mainsnak": {"snaktype": "value", "property": "P910", "hash": "839cb661dd8b62a80799548e18b90db527de24cd", "datavalue": {"value": {"entity-type": "item", "numeric-id": 13299250, "id": "Q13299250"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q10541491$8c41ac7d-463c-0da7-d4c4-8ad19ba2ca2b", "rank": "normal"}], "P1889": [{"mainsnak": {"snaktype": "value", "property": "P1889", "hash": "d387d0b27e7cb3ce4f0e52a27eea1f8e71a02a5c", "datavalue": {"value": {"entity-type": "item", "numeric-id": 1938208, "id": "Q1938208"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q10541491$1b4fb126-41f7-f6c2-91d8-21077e430a35", "rank": "normal"}], "P1269": [{"mainsnak": {"snaktype": "value", "property": "P1269", "hash": "40e6b2c05fca59667f7598249ce704f8f0dfa4f9", "datavalue": {"value": {"entity-type": "item", "numeric-id": 3778211, "id": "Q3778211"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q10541491$69059f98-4e79-4a11-43f9-ffa7377f29d3", "rank": "normal"}]}}}, "success": 1}
