+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q106559804.json
@@ -1,0 +1,1 @@
+{"entities": {"Q106559804": {"pageid": 101845301, "ns": 0, "title": "Q106559804", "lastrevid": 2319404162, "modified": "2025-03-03T14:50:40Z", "type": "item", "id": "Q106559804", "labels": {"en": {"language": "en", "value": "person or organization"}}, "descriptions": {"en": {"language": "en", "value": "class of agents"}}, "aliases": {"en": [{"language": "en", "value": "person or organisation"}, {"language": "en", "value": "organisation or person"}, {"language": "en", "value": "organization or person"}, {"language": "en", "value": "human or organization"}]}, "claims": {"P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "eb6887f11ff6198dff8aeba1410546badb442892", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24229398, "id": "Q24229398"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q106559804$e24e2e82-40f3-fd95-8bce-7683f53101f3", "rank": "normal"}], "P1269": [{"mainsnak": {"snaktype": "value", "property": "P1269", "hash": "40e6b2c05fca59667f7598249ce704f8f0dfa4f9", "datavalue": {"value": {"entity-type": "item", "numeric-id": 3778211, "id": "Q3778211"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q106559804$39b25b51-4a20-5d92-e768-4b21e8c82240", "rank": "normal"}], "P1382": [{"mainsnak": {"snaktype": "value", "property": "P1382", "hash": "2c839757b0efd426d77c14f62781812bf44c2d4c", "datavalue": {"value": {"entity-type": "item", "numeric-id": 3778211, "id": "Q3778211"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q106559804$eaa29815-49ac-a214-b5db-935e54e8c38a", "rank": "normal"}], "P2738": [{"mainsnak": {"snaktype": "value", "property": "P2738", "hash": "4fa7d4837acece3e3ee87aa78037ef4527696ea4", "datavalue": {"value": {"entity-type": "item", "numeric-id": 23766486, "id": "Q23766486"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P11260": [{"snaktype": "value", "property": "P11260", "hash": "a0134eeae8dc5577fdf4207a72bb34c2aacf3f4e", "datavalue": {"value": {"entity-type": "item", "numeric-id": 43229, "id": "Q43229"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, {"snaktype": "value", "property": "P11260", "hash": "8e40d6ce51d102c2d1da0f67fd754d04aa8d5fdb", "datavalue": {"value": {"entity-type": "item", "numeric-id": 215627, "id": "Q215627"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P11260"], "id": "Q106559804$b7768fc6-4455-69ee-3992-b669aef01eec", "rank": "normal"}]}}}, "success": 1}
