+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q10855152.json
@@ -1,0 +1,1 @@
+{"entities": {"Q10855152": {"pageid": 12127392, "ns": 0, "title": "Q10855152", "lastrevid": 2296662350, "modified": "2025-01-10T09:48:59Z", "type": "item", "id": "Q10855152", "labels": {"en": {"language": "en", "value": "individual organism"}}, "descriptions": {"en": {"language": "en", "value": "individual living thing"}}, "aliases": {"en": [{"language": "en", "value": "individual being"}, {"language": "en", "value": "individual of an organism"}, {"language": "en", "value": "individual life"}]}, "claims": {"P646": [{"mainsnak": {"snaktype": "value", "property": "P646", "hash": "11a82df05a2045ee13db9727f24a5ff57d431e83", "datavalue": {"value": "/m/0nbhywv", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q10855152$202DEF6E-5744-4668-9665-401678877273", "rank": "normal"}], "P11408": [{"mainsnak": {"snaktype": "value", "property": "P11408", "hash": "eb37cb0dc3195b7c5bcf246ec3d5835e3675c111", "datavalue": {"value": "\u500b\u4f53", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q10855152$b64d2f75-bb66-46d7-bea6-07d17cbf768c", "rank": "normal"}], "P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "d38a6e64e83cc318aeb610a3677ea2521ac0e31b", "datavalue": {"value": {"entity-type": "item", "numeric-id": 795052, "id": "Q795052"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q10855152$c1f3db2f-4947-d5d1-c0b1-055b0acd14c6", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P279", "hash": "cc1eb50276462e05cca1d1bb823266f6cc24073c", "datavalue": {"value": {"entity-type": "item", "numeric-id": 7239, "id": "Q7239"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q10855152$12cb1f82-4758-1762-f52b-8906882f4960", "rank": "normal"}], "P1963": [{"mainsnak": {"snaktype": "value", "property": "P1963", "hash": "cf74ee35f9a696c4345504c2121822533335355e", "datavalue": {"value": {"entity-type": "property", "numeric-id": 10241, "id": "P10241"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}, "type": "statement", "id": "Q10855152$1DC61235-B991-4A87-9B20-C3B4C8A60DB9", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P1963", "hash": "be6d5b6e6a3a8c6e7ad2166fbb08237efaf27b52", "datavalue": {"value": {"entity-type": "property", "numeric-id": 21, "id": "P21"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}, "type": "statement", "id": "Q10855152$495249E2-F4F0-4039-BE01-E5F4226E3074", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P1963", "hash": "414816073ce1c46e03f1812b3be7dad3b897c164", "datavalue": {"value": {"entity-type": "property", "numeric-id": 127, "id": "P127"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}, "type": "statement", "id": "Q10855152$67773F7B-218C-433E-A580-A6DFEE1435D0", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P1963", "hash": "5dbfb9597f2bce5bba4ba8ca5054fa1a9f86201e", "datavalue": {"value": {"entity-type": "property", "numeric-id": 1449, "id": "P1449"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}, "type": "statement", "id": "Q10855152$DBC4BA96-3B08-4A52-9DDA-18E17256F40E", "rank": "normal"}], "P1687": [{"mainsnak": {"snaktype": "value", "property": "P1687", "hash": "eefa154aa290cefd1c09508e9cb2cd4bd46fc92d", "datavalue": {"value": {"entity-type": "property", "numeric-id": 10241, "id": "P10241"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}, "type": "statement", "id": "Q10855152$1110CF1F-EA05-48BC-97E4-543D7CB0AFBC", "rank": "normal"}], "P910": [{"mainsnak": {"snaktype": "value", "property": "P910", "hash": "d01473101a0e0d8caf786705515ff80416475fac", "datavalue": {"value": {"entity-type": "item", "numeric-id": 8546870, "id": "Q8546870"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q10855152$3AEDEE05-647A-4907-98F2-4DDE257720CC", "rank": "normal"}], "P460": [{"mainsnak": {"snaktype": "value", "property": "P460", "hash": "799824c5cdd59788d2e898b7a5e1f343d0aa6559", "datavalue": {"value": {"entity-type": "item", "numeric-id": 12539981, "id": "Q12539981"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q10855152$9E8734EA-8717-4A10-9048-FB288AD73524", "rank": "normal"}]}}}, "success": 1}
