+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q115372263.json
@@ -1,0 +1,1 @@
+{"entities": {"Q115372263": {"pageid": 109956584, "ns": 0, "title": "Q115372263", "lastrevid": 2296221224, "modified": "2025-01-08T15:06:23Z", "type": "item", "id": "Q115372263", "labels": {"en": {"language": "en", "value": "word or phrase"}}, "descriptions": {"en": {"language": "en", "value": "sequence of one or more words"}}, "aliases": {"en": [{"language": "en", "value": "phrase"}, {"language": "en", "value": "word and phrase"}]}, "claims": {"P361": [{"mainsnak": {"snaktype": "value", "property": "P361", "hash": "612b2f31f3853379ac60819f11daa63185765371", "datavalue": {"value": {"entity-type": "item", "numeric-id": 41796, "id": "Q41796"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q115372263$ee4e890d-4488-61e0-d48e-7b654f013c2a", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P361", "hash": "b79af78800b49062efdf9619db6847478b21ba93", "datavalue": {"value": {"entity-type": "item", "numeric-id": 117364, "id": "Q117364"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q115372263$b49fa876-4731-87f9-70ff-4c5dc61a512f", "rank": "normal"}], "P2738": [{"mainsnak": {"snaktype": "value", "property": "P2738", "hash": "4fa7d4837acece3e3ee87aa78037ef4527696ea4", "datavalue": {"value": {"entity-type": "item", "numeric-id": 23766486, "id": "Q23766486"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P11260": [{"snaktype": "value", "property": "P11260", "hash": "02763febd4b5f7fa6982e1b9c9557fe401f492ee", "datavalue": {"value": {"entity-type": "item", "numeric-id": 8171, "id": "Q8171"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, {"snaktype": "value", "property": "P11260", "hash": "534356eee7fd9836dbfe232c0d9db073de2cfc24", "datavalue": {"value": {"entity-type": "item", "numeric-id": 187931, "id": "Q187931"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P11260"], "id": "Q115372263$17f9acd8-423e-d843-241f-7863b21475a1", "rank": "normal"}], "P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "71506ebf13d3857fc73d2cba0d67a591c268bb26", "datavalue": {"value": {"entity-type": "item", "numeric-id": 115786086, "id": "Q115786086"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q115372263$69fe685b-49f9-d44c-a0b1-94dc409544cc", "rank": "normal", "references": [{"hash": "607086bdcf30e89146be969b498d3557f5f40100", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "4fb35a32f6f6c54b6769d78a80594e3507a49f60", "datavalue": {"value": "https://www.ahdictionary.com/word/search.html?q=linguistic%20form", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P854"]}]}], "P1476": [{"mainsnak": {"snaktype": "value", "property": "P1476", "hash": "f8c8be29974a1d2dc8802dc150349a264207b58b", "datavalue": {"value": {"text": "phrase", "language": "en"}, "type": "monolingualtext"}, "datatype": "monolingualtext"}, "type": "statement", "id": "Q115372263$4bf4c064-4a5d-15c2-c922-158e38c1f76f", "rank": "normal", "references": [{"hash": "6232dee44a772126a12f1f9b352ba99568dab4f4", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "3ef771955ee395483ebe064093ab8ffcd35e85ab", "datavalue": {"value": "https://www.merriam-webster.com/dictionary/phrase", "type": "string"}, "datatype": "url"}], "P1476": [{"snaktype": "value", "property": "P1476", "hash": "dfea5e076ddb136e9d86ba373710f91fd7e3e845", "datavalue": {"value": {"text": "sense 4 allows 1 or more words", "language": "en"}, "type": "monolingualtext"}, "datatype": "monolingualtext"}]}, "snaks-order": ["P854", "P1476"]}, {"hash": "eead237f903809d5d67e42d1702c736b634ceaca", "snaks": {"P8814": [{"snaktype": "value", "property": "P8814", "hash": "0d1ad5aeb34fcab3817dad4200839ddc39b11572", "datavalue": {"value": "06327041-n", "type": "string"}, "datatype": "external-id"}]}, "snaks-order": ["P8814"]}]}], "P2561": [{"mainsnak": {"snaktype": "value", "property": "P2561", "hash": "776c994591a65625fd712c33680c643b6b368c1a", "datavalue": {"value": {"text": "phrase", "language": "en"}, "type": "monolingualtext"}, "datatype": "monolingualtext"}, "type": "statement", "id": "Q115372263$d78ced9f-4988-3d19-46d0-21258ece7bad", "rank": "normal", "references": [{"hash": "7892c647b3af3e649b5e7a4aa7f83b715d7dca20", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "3ef771955ee395483ebe064093ab8ffcd35e85ab", "datavalue": {"value": "https://www.merriam-webster.com/dictionary/phrase", "type": "string"}, "datatype": "url"}], "P2315": [{"snaktype": "value", "property": "P2315", "hash": "71774b3435f8f5ad3fc2bb834234de3639f7b721", "datavalue": {"value": {"text": "sense 4 of MW allows 1 or more words", "language": "en"}, "type": "monolingualtext"}, "datatype": "monolingualtext"}]}, "snaks-order": ["P854", "P2315"]}, {"hash": "eead237f903809d5d67e42d1702c736b634ceaca", "snaks": {"P8814": [{"snaktype": "value", "property": "P8814", "hash": "0d1ad5aeb34fcab3817dad4200839ddc39b11572", "datavalue": {"value": "06327041-n", "type": "string"}, "datatype": "external-id"}]}, "snaks-order": ["P8814"]}]}], "P910": [{"mainsnak": {"snaktype": "value", "property": "P910", "hash": "2d319f106935637595181cf70d20f90bed354604", "datavalue": {"value": {"entity-type": "item", "numeric-id": 5641890, "id": "Q5641890"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q115372263$f70158e5-4fc0-7489-455e-a435224f8a7c", "rank": "normal"}]}}}, "success": 1}
