+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q15916540.json
@@ -1,0 +1,1 @@
+{"entities": {"Q15916540": {"pageid": 17541190, "ns": 0, "title": "Q15916540", "lastrevid": 2296436685, "modified": "2025-01-09T11:45:45Z", "type": "item", "id": "Q15916540", "labels": {"en": {"language": "en", "value": "semantic unit"}}, "descriptions": {"en": {"language": "en", "value": "linguistic unit that carries meaning (citation needed; does it really *carry* meaning? how many meanings does a semantic unit have?)"}}, "aliases": {}, "claims": {"P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "a140e0283d5f8ffeed0ae033d9a1ab930bb8e712", "datavalue": {"value": {"entity-type": "item", "numeric-id": 1786828, "id": "Q1786828"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P1013": [{"snaktype": "value", "property": "P1013", "hash": "73e1d1f0a166cc1e0a8d7fa02a10d954891b6f25", "datavalue": {"value": {"entity-type": "item", "numeric-id": 183046, "id": "Q183046"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P1013"], "id": "Q15916540$3bbc3354-4c42-7aba-ebd9-6e3cd0a22980", "rank": "normal"}], "P910": [{"mainsnak": {"snaktype": "value", "property": "P910", "hash": "d4c7586ea36cf499aadac0527e381e2abd539f14", "datavalue": {"value": {"entity-type": "item", "numeric-id": 7210931, "id": "Q7210931"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q15916540$8be5a8cf-4867-2f81-ec5d-677b9482701c", "rank": "normal"}], "P8408": [{"mainsnak": {"snaktype": "value", "property": "P8408", "hash": "59b0a1b08d5e68af598fe210228d8a31b21d019d", "datavalue": {"value": "SemanticUnit", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q15916540$7F805FCF-5296-4485-9367-95B7B8F09406", "rank": "normal", "references": [{"hash": "9a681f9dd95c90224547c404e11295f4f7dcf54e", "snaks": {"P248": [{"snaktype": "value", "property": "P248", "hash": "9d5780dddffa8746637a9929a936ab6b0f601e24", "datavalue": {"value": {"entity-type": "item", "numeric-id": 64139102, "id": "Q64139102"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}], "P813": [{"snaktype": "value", "property": "P813", "hash": "622a5a27fa5b25e7e7984974e9db494cf8460990", "datavalue": {"value": {"time": "+2020-07-09T00:00:00Z", "timezone": 0, "before": 0, "after": 0, "precision": 11, "calendarmodel": "http://www.wikidata.org/entity/Q1985727"}, "type": "time"}, "datatype": "time"}]}, "snaks-order": ["P248", "P813"]}]}], "P1535": [{"mainsnak": {"snaktype": "value", "property": "P1535", "hash": "a2bcf0a35e14d654b4614b9f50d3dfc42921d5c5", "datavalue": {"value": {"entity-type": "item", "numeric-id": 1066689, "id": "Q1066689"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q15916540$e59ab773-4c03-12ac-c28b-914833dfc9d5", "rank": "normal"}], "P361": [{"mainsnak": {"snaktype": "value", "property": "P361", "hash": "e67e1c0456f98bf6afc3a9913e9bf597ff2cc8f6", "datavalue": {"value": {"entity-type": "item", "numeric-id": 1066689, "id": "Q1066689"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q15916540$779255ca-4450-6f7d-12c4-b0a080fdae85", "rank": "normal"}], "P1552": [{"mainsnak": {"snaktype": "value", "property": "P1552", "hash": "c5d90e1ba7a716f1e61ba6198d75453cd31711d1", "datavalue": {"value": {"entity-type": "item", "numeric-id": 21652121, "id": "Q21652121"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P642": [{"snaktype": "value", "property": "P642", "hash": "fd000607ba823a23462e5c929859cda88b0ef346", "datavalue": {"value": {"entity-type": "item", "numeric-id": 234460, "id": "Q234460"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P642"], "id": "Q15916540$4072bb24-47a3-3d69-5674-273123e26efb", "rank": "normal", "references": [{"hash": "ea9ba8a4e68751eaceb60b34065a92aba16144bf", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "cafb133751b2a8895ad8a424b1cfd864dba35498", "datavalue": {"value": "https://books.google.com/books?id=ZxbUFm3aRkUC&pg=PA853", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P854"]}]}, {"mainsnak": {"snaktype": "value", "property": "P1552", "hash": "41853f675789a52d86ada498e53455dfbafd2808", "datavalue": {"value": {"entity-type": "item", "numeric-id": 183046, "id": "Q183046"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q15916540$71e35b19-476a-0388-02b2-44b515438498", "rank": "normal"}], "P2315": [{"mainsnak": {"snaktype": "value", "property": "P2315", "hash": "138a467578c7e6e112b8e9548097c36d47aaa123", "datavalue": {"value": {"text": "example: even though", "language": "en"}, "type": "monolingualtext"}, "datatype": "monolingualtext"}, "type": "statement", "id": "Q15916540$b45d88fe-4602-005a-693b-a989020d0333", "rank": "normal", "references": [{"hash": "715c390933cbee288414ff972bfc1706d63df6f3", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "4a9de2dc75faf43fa6d2e30e9f846a78acb4ebc7", "datavalue": {"value": "https://books.google.com/books?id=O_iCDwAAQBAJ&pg=PT91", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P854"]}]}, {"mainsnak": {"snaktype": "value", "property": "P2315", "hash": "aa0fabc76b21463314910940bd476843de75a88b", "datavalue": {"value": {"text": "counter-example: did several", "language": "en"}, "type": "monolingualtext"}, "datatype": "monolingualtext"}, "type": "statement", "id": "Q15916540$3da5959d-4c7d-b562-0a82-f09ce18366e0", "rank": "normal", "references": [{"hash": "715c390933cbee288414ff972bfc1706d63df6f3", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "4a9de2dc75faf43fa6d2e30e9f846a78acb4ebc7", "datavalue": {"value": "https://books.google.com/books?id=O_iCDwAAQBAJ&pg=PT91", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P854"]}]}, {"mainsnak": {"snaktype": "value", "property": "P2315", "hash": "37db8c5561f4a43c7b6048ea51c91e85f065a025", "datavalue": {"value": {"text": "example: \"chair (Semantic Unit)\"", "language": "en"}, "type": "monolingualtext"}, "datatype": "monolingualtext"}, "type": "statement", "id": "Q15916540$cc5d805c-4fec-fd90-895a-4c3c0c16d05d", "rank": "normal", "references": [{"hash": "8c40886792973cd6bd3a1d4f03bc691bbdb3fb4c", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "190696575983ebb3e5039cb067ee44d60228cc24", "datavalue": {"value": "https://books.google.com/books?id=IfUgEAAAQBAJ&pg=PT159", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P854"]}]}, {"mainsnak": {"snaktype": "value", "property": "P2315", "hash": "926a8911f84698da80a70d2dde0ceb21e4325c46", "datavalue": {"value": {"text": "example: disclose", "language": "en"}, "type": "monolingualtext"}, "datatype": "monolingualtext"}, "type": "statement", "id": "Q15916540$c953b452-4cb4-bf78-e409-a4da54b7bb35", "rank": "normal", "references": [{"hash": "904d618590755b00a39d3ea35d866fac5921e079", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "fc1a648ceda9d2ebea65a07029d9cf2bb5f501ac", "datavalue": {"value": "https://books.google.com/books?id=S_lRAXSpymkC&pg=PA67", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P854"]}]}, {"mainsnak": {"snaktype": "value", "property": "P2315", "hash": "a8519c61e7a7302cde74e0f57492474690cce263", "datavalue": {"value": {"text": "example: restaurant", "language": "en"}, "type": "monolingualtext"}, "datatype": "monolingualtext"}, "type": "statement", "id": "Q15916540$54c86bbf-41df-610b-b156-6049bde5d47a", "rank": "normal", "references": [{"hash": "7264b9cbe01a885d009cfc16c2c4f9b4a08d1d1c", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "5afef771d70838ac9afc4869013f50d44c314c0e", "datavalue": {"value": "https://books.google.com/books?id=SQc31J3Yyn4C&pg=PA54", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P854"]}]}]}}}, "success": 1}
