+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q19868531.json
@@ -1,0 +1,1 @@
+{"entities": {"Q19868531": {"pageid": 21497572, "ns": 0, "title": "Q19868531", "lastrevid": 2321951899, "modified": "2025-03-09T02:20:48Z", "type": "item", "id": "Q19868531", "labels": {"en": {"language": "en", "value": "formal ontology concept"}}, "descriptions": {"en": {"language": "en", "value": "formally defined concept in the context of an ontology"}}, "aliases": {}, "claims": {"P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "978e940babbe8519762ba8269076a11498b25980", "datavalue": {"value": {"entity-type": "item", "numeric-id": 151885, "id": "Q151885"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q19868531$af801cce-4351-c300-294e-1431de660229", "rank": "normal"}], "P361": [{"mainsnak": {"snaktype": "value", "property": "P361", "hash": "835ceaa54f64afacde339011973c19ef6cbabfa6", "datavalue": {"value": {"entity-type": "item", "numeric-id": 324254, "id": "Q324254"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q19868531$31800870-449f-041d-89dc-eb5b7719adbe", "rank": "normal"}]}}}, "success": 1}
