+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q2145290.json
@@ -1,0 +1,1 @@
+{"entities": {"Q2145290": {"pageid": 2070986, "ns": 0, "title": "Q2145290", "lastrevid": 2320000036, "modified": "2025-03-04T20:11:17Z", "type": "item", "id": "Q2145290", "labels": {"en": {"language": "en", "value": "mental representation"}}, "descriptions": {"en": {"language": "en", "value": "hypothetical internal cognitive symbol that represents external reality"}}, "aliases": {"en": [{"language": "en", "value": "cognitive representation"}]}, "claims": {"P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "62545af5a446d3ca93c7796ddc9d92708360d8e4", "datavalue": {"value": {"entity-type": "item", "numeric-id": 1272626, "id": "Q1272626"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q2145290$7bb377fa-476a-a265-495b-5152beb47abd", "rank": "normal"}], "P2579": [{"mainsnak": {"snaktype": "value", "property": "P2579", "hash": "7375e0aca7677571a994050a1ec9ef7a3161f299", "datavalue": {"value": {"entity-type": "item", "numeric-id": 147638, "id": "Q147638"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q2145290$9d48769a-4c6a-dfce-f2cf-08ef736b1135", "rank": "normal"}], "P244": [{"mainsnak": {"snaktype": "value", "property": "P244", "hash": "ad6ceb4f3c745c27bb358f36c9121042ddbac7d4", "datavalue": {"value": "sh88004829", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q2145290$c5fc71d3-49bc-44eb-ecda-b91423fe8adb", "rank": "normal"}], "P268": [{"mainsnak": {"snaktype": "value", "property": "P268", "hash": "ccd7b488046db3ff58776d4d02e23e158bb72248", "datavalue": {"value": "12315800k", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q2145290$915daaad-40e8-6d13-0ce7-5a320ec489ce", "rank": "normal"}], "P227": [{"mainsnak": {"snaktype": "value", "property": "P227", "hash": "da90dfa72cb82237b60c11e9349bc6948fa8d95f", "datavalue": {"value": "4049534-6", "type": "string"}, "datatype": "external-id"}, "type": "statement", "qualifiers": {"P1810": [{"snaktype": "value", "property": "P1810", "hash": "258345c41ad7d958965b69d5f6399a42f2043461", "datavalue": {"value": "Wissensrepr\u00e4sentation", "type": "string"}, "datatype": "string"}], "P4970": [{"snaktype": "value", "property": "P4970", "hash": "273cae626593c807c9644a316af370c28617753b", "datavalue": {"value": "Repr\u00e4sentation <Psychologie>", "type": "string"}, "datatype": "string"}], "P4070": [{"snaktype": "value", "property": "P4070", "hash": "a29d2fce83454522e7eb3f95e2ab7f163527824c", "datavalue": {"value": {"entity-type": "item", "numeric-id": 3478658, "id": "Q3478658"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P1810", "P4970", "P4070"], "id": "Q2145290$2c1a0193-4cae-8758-5096-c0547b6b42fa", "rank": "normal", "references": [{"hash": "424a9bef05afca53f5ce38674f43729d71c1e6ca", "snaks": {"P813": [{"snaktype": "value", "property": "P813", "hash": "1894a34c5efbd49656522feda11beecd33129019", "datavalue": {"value": {"time": "+2023-05-21T00:00:00Z", "timezone": 0, "before": 0, "after": 0, "precision": 11, "calendarmodel": "http://www.wikidata.org/entity/Q1985727"}, "type": "time"}, "datatype": "time"}]}, "snaks-order": ["P813"]}]}], "P361": [{"mainsnak": {"snaktype": "value", "property": "P361", "hash": "9577875b53cff629b7704855d228090295dd2275", "datavalue": {"value": {"entity-type": "item", "numeric-id": 77468620, "id": "Q77468620"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q2145290$C7500099-1CC8-41B1-B5AF-3C2172135B28", "rank": "normal"}], "P6366": [{"mainsnak": {"snaktype": "value", "property": "P6366", "hash": "7567446b046dbed0f8674ab48c97f89f9b180ce9", "datavalue": {"value": "96199812", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q2145290$765468B4-A848-4992-86DA-19BB2B60730C", "rank": "normal"}], "P2347": [{"mainsnak": {"snaktype": "value", "property": "P2347", "hash": "dd92c371f72274d5b5b24792ace4f1656b3adb18", "datavalue": {"value": "1407", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q2145290$04828B14-6AAA-4EF4-BF75-E2EF7046F5BF", "rank": "normal", "references": [{"hash": "f9bf1a1f034ddd51bd9928ac535e0f57d748e2cf", "snaks": {"P248": [{"snaktype": "value", "property": "P248", "hash": "7133f11674741f52cadaae6029068fad9cbb52e3", "datavalue": {"value": {"entity-type": "item", "numeric-id": 89345680, "id": "Q89345680"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "snaks-order": ["P248"]}]}], "P646": [{"mainsnak": {"snaktype": "value", "property": "P646", "hash": "aaac094288d0dbb5336151329ea006171c29b771", "datavalue": {"value": "/m/064r8p6", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q2145290$9FC5F5DA-7BED-4D60-B97E-A820B7A2E606", "rank": "normal"}], "P10283": [{"mainsnak": {"snaktype": "value", "property": "P10283", "hash": "56d1f535c0e6fc8048d4fa8400fbdf0fd1057638", "datavalue": {"value": "C96199812", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q2145290$E001DE74-5CF0-49E9-A8D1-071692686DE5", "rank": "normal", "references": [{"hash": "c13ba4d7902ead2693f79396bc25bd35d4d14030", "snaks": {"P248": [{"snaktype": "value", "property": "P248", "hash": "4a4f26a5361b5707266e48e425bf2be2f99fd2ab", "datavalue": {"value": {"entity-type": "item", "numeric-id": 107507571, "id": "Q107507571"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}], "P813": [{"snaktype": "value", "property": "P813", "hash": "435834d08182bb9f3dbe974ba9840af0f12899cc", "datavalue": {"value": {"time": "+2022-01-26T00:00:00Z", "timezone": 0, "before": 0, "after": 0, "precision": 11, "calendarmodel": "http://www.wikidata.org/entity/Q1985727"}, "type": "time"}, "datatype": "time"}], "P854": [{"snaktype": "value", "property": "P854", "hash": "a4a4bf53f22268815c51ec10fed608da703c9c7f", "datavalue": {"value": "https://docs.openalex.org/download-snapshot/snapshot-data-format", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P248", "P813", "P854"]}]}], "P460": [{"mainsnak": {"snaktype": "value", "property": "P460", "hash": "073773f112a24a905798d80c035a3ebabf6d0ba5", "datavalue": {"value": {"entity-type": "item", "numeric-id": 112040407, "id": "Q112040407"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q2145290$46d3f63d-40e2-98f3-4dbf-5e19f756f021", "rank": "normal"}], "P8189": [{"mainsnak": {"snaktype": "value", "property": "P8189", "hash": "cf5de401c1f563ee264bb078406a4f8935c2905b", "datavalue": {"value": "987007548727805171", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q2145290$A80047E4-6F25-4E6F-B9F8-009F58784724", "rank": "normal", "references": [{"hash": "ecb7a39cf436eb1fba419853af168ef9f2ba1dab", "snaks": {"P248": [{"snaktype": "value", "property": "P248", "hash": "a03bb6ffe37be9554492f4eea5bd18afcb57df19", "datavalue": {"value": {"entity-type": "item", "numeric-id": 106509962, "id": "Q106509962"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "snaks-order": ["P248"]}]}], "P10782": [{"mainsnak": {"snaktype": "value", "property": "P10782", "hash": "0fcf2a51ba641ddfcc0f2014500f71748d088c44", "datavalue": {"value": "328", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q2145290$287F057E-97AC-4128-9016-C9577C43E4B0", "rank": "normal"}]}}}, "success": 1}
