+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q21871294.json
@@ -1,0 +1,1 @@
+{"entities": {"Q21871294": {"pageid": 23908138, "ns": 0, "title": "Q21871294", "lastrevid": 2296924070, "modified": "2025-01-11T09:05:16Z", "type": "item", "id": "Q21871294", "labels": {"en": {"language": "en", "value": "group or class of organisms"}}, "descriptions": {"en": {"language": "en", "value": "group of organisms"}}, "aliases": {"en": [{"language": "en", "value": "group of organisms"}, {"language": "en", "value": "class of organisms"}, {"language": "en", "value": "class of living things"}, {"language": "en", "value": "living organism class"}, {"language": "en", "value": "organism class"}, {"language": "en", "value": "group or class of living organisms"}]}, "claims": {"P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "33c0fd1431aa0fb276f87bb51b4a956d9440bb15", "datavalue": {"value": {"entity-type": "item", "numeric-id": 16889133, "id": "Q16889133"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P642": [{"snaktype": "value", "property": "P642", "hash": "4c012de4364fa25db13a3cb04174b8a1539fba5f", "datavalue": {"value": {"entity-type": "item", "numeric-id": 7239, "id": "Q7239"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P642"], "id": "Q21871294$2dcf69c4-49b3-fd5c-63fa-48c23d7dc669", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P279", "hash": "e5fbc99ed9198dd6b444b294471426c62b167830", "datavalue": {"value": {"entity-type": "item", "numeric-id": 98119401, "id": "Q98119401"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q21871294$b3fc4a61-46a5-6367-d977-1662de4ff5bf", "rank": "normal"}], "P31": [{"mainsnak": {"snaktype": "value", "property": "P31", "hash": "a7e32719e3a70a25fce867c2ad70c306d3521fc3", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24017414, "id": "Q24017414"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P8225": [{"snaktype": "value", "property": "P8225", "hash": "c42b95c1071479a6438d6fc54b529596a9c24bd1", "datavalue": {"value": {"entity-type": "item", "numeric-id": 7239, "id": "Q7239"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P8225"], "id": "Q21871294$c0154b16-46ab-a768-6a43-736a7bc68bd8", "rank": "normal"}], "P8225": [{"mainsnak": {"snaktype": "value", "property": "P8225", "hash": "c42b95c1071479a6438d6fc54b529596a9c24bd1", "datavalue": {"value": {"entity-type": "item", "numeric-id": 7239, "id": "Q7239"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q21871294$f378d883-40ad-6d85-62a0-369525e709fa", "rank": "normal"}], "P1889": [{"mainsnak": {"snaktype": "value", "property": "P1889", "hash": "509333c079e870a7c2d784d476c7aebafd27b01c", "datavalue": {"value": {"entity-type": "item", "numeric-id": 16334298, "id": "Q16334298"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q21871294$243e8a2a-4aeb-f927-1719-4151a97fa0ea", "rank": "normal"}], "P1382": [{"mainsnak": {"snaktype": "value", "property": "P1382", "hash": "0d18369672cb8024ca55a584711653170b649328", "datavalue": {"value": {"entity-type": "item", "numeric-id": 113657383, "id": "Q113657383"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q21871294$bd5fe884-4dfd-3faa-1ce5-a5f414570981", "rank": "normal"}]}}}, "success": 1}
