+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q2198779.json
@@ -1,0 +1,1 @@
+{"entities": {"Q2198779": {"pageid": 2123087, "ns": 0, "title": "Q2198779", "lastrevid": 2303853209, "modified": "2025-01-29T16:08:14Z", "type": "item", "id": "Q2198779", "labels": {"en": {"language": "en", "value": "unit"}}, "descriptions": {"en": {"language": "en", "value": "entity regarded or used as an elementary structural or functional constituent to measure, analyse or describe another entity"}}, "aliases": {}, "claims": {"P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "6aedaf1acb8b0a235981d0a5e0c857177235f0ec", "datavalue": {"value": {"entity-type": "item", "numeric-id": 7048977, "id": "Q7048977"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q2198779$7f0d6d08-4439-67ed-cce8-6f79c1439c46", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P279", "hash": "a140e0283d5f8ffeed0ae033d9a1ab930bb8e712", "datavalue": {"value": {"entity-type": "item", "numeric-id": 1786828, "id": "Q1786828"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P1932": [{"snaktype": "value", "property": "P1932", "hash": "abdda5d630c0b93c706b05087f5dbae73629f0be", "datavalue": {"value": "constituent", "type": "string"}, "datatype": "string"}]}, "qualifiers-order": ["P1932"], "id": "Q2198779$ac2836e2-49ec-d3d9-44e3-b08ce69576f3", "rank": "deprecated", "references": [{"hash": "cb85a8ff63e4a8fba1d06e8ebc99ffbcc13180df", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "28acedc0c38daa1ba38ef9fc632c0c366ee01b33", "datavalue": {"value": "https://www.merriam-webster.com/dictionary/unit", "type": "string"}, "datatype": "url"}], "P2315": [{"snaktype": "value", "property": "P2315", "hash": "b30749b277f51d1e8854f3044c3ace5808604f08", "datavalue": {"value": {"text": "3a", "language": "en"}, "type": "monolingualtext"}, "datatype": "monolingualtext"}]}, "snaks-order": ["P854", "P2315"]}, {"hash": "e75894425a8a76dc0b34cc655a6650b272a98b90", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "6d979f44082dca840d26d1943dd665d3ba4388c8", "datavalue": {"value": "https://www.ahdictionary.com/word/search.html?q=unit", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P854"]}]}, {"mainsnak": {"snaktype": "value", "property": "P279", "hash": "be9881aa5dd168f3caedefdb7c8dd6dc0596174f", "datavalue": {"value": {"entity-type": "item", "numeric-id": 47574, "id": "Q47574"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P2241": [{"snaktype": "value", "property": "P2241", "hash": "cbae4cc270287a48939ecce7e697f4cc27b9dac5", "datavalue": {"value": {"entity-type": "item", "numeric-id": 118134102, "id": "Q118134102"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P2241"], "id": "Q2198779$b3bcaecc-4fc5-0168-558c-161e96fbd899", "rank": "deprecated"}], "P2671": [{"mainsnak": {"snaktype": "value", "property": "P2671", "hash": "27961dad49e2c87d49dfc6dc6ec9354cd4182cd4", "datavalue": {"value": "/g/120zclw4", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q2198779$235F34C8-7899-44F2-A987-BC8A5D5E7951", "rank": "normal"}], "P8313": [{"mainsnak": {"snaktype": "value", "property": "P8313", "hash": "e15ad8aa4e3ec9027f93246a00c32c9912429435", "datavalue": {"value": "enhed", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q2198779$F4DC7482-4385-4457-BF96-918B0817FF75", "rank": "normal"}], "P1552": [{"mainsnak": {"snaktype": "value", "property": "P1552", "hash": "30949c0fabb3eb997adbad176328c2a8dc0b136a", "datavalue": {"value": {"entity-type": "item", "numeric-id": 199, "id": "Q199"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P1932": [{"snaktype": "value", "property": "P1932", "hash": "3c369c0a80564cf2ff24daf6160f9e8017d922c9", "datavalue": {"value": "unity", "type": "string"}, "datatype": "string"}], "P1480": [{"snaktype": "value", "property": "P1480", "hash": "87c54e32233b91cef34a47b0c6cd09ce5c6a84a4", "datavalue": {"value": {"entity-type": "item", "numeric-id": 41719, "id": "Q41719"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P1932", "P1480"], "id": "Q2198779$f6382964-4450-c7fd-f860-980b56824478", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P1552", "hash": "654e62c0842223b9264d10eb6003fa2d03ce7ace", "datavalue": {"value": {"entity-type": "item", "numeric-id": 19750549, "id": "Q19750549"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q2198779$2eb55424-4c99-92ed-f84e-0cd6012f5d3d", "rank": "normal", "references": [{"hash": "75448f4e6a6f4758786a05626d156798358c9923", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "61370158ad88fb211886e5c311041315679b2b82", "datavalue": {"value": "https://www.oxfordlearnersdictionaries.com/us/definition/english/unit", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P854"]}]}], "P361": [{"mainsnak": {"snaktype": "value", "property": "P361", "hash": "cc499f22ef356b14a42e1a5c3057f25d47737541", "datavalue": {"value": {"entity-type": "item", "numeric-id": 16868672, "id": "Q16868672"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q2198779$cb71e044-4aed-cc80-09cb-3cc30b92e12c", "rank": "normal", "references": [{"hash": "190c06ca337f0ff179dcca2e9c5535347f4f4245", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "f4895d74ba98ab8764111717e4ce93bc1ff926f6", "datavalue": {"value": "https://www.macmillandictionary.com/dictionary/american/unit", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P854"]}, {"hash": "92a499d53c6a90e46fcb83b6aadd58b021915c8c", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "28acedc0c38daa1ba38ef9fc632c0c366ee01b33", "datavalue": {"value": "https://www.merriam-webster.com/dictionary/unit", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P854"]}, {"hash": "407f58d3285d4d4dcae97b692dd487ae7a2ea808", "snaks": {"P854": [{"snaktype": "value", "property": "P854", "hash": "40320e72690c46b4cc318f3e7afcbd3064a843a5", "datavalue": {"value": "https://dictionary.cambridge.org/dictionary/english/unit?q=unit_1", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P854"]}]}], "P460": [{"mainsnak": {"snaktype": "value", "property": "P460", "hash": "12e40c4c449ea8eeae56c37163222dfc1c02a0c7", "datavalue": {"value": {"entity-type": "item", "numeric-id": 78754808, "id": "Q78754808"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q2198779$76657894-4d99-e9c0-c3cb-2778a02a4914", "rank": "normal"}], "P1889": [{"mainsnak": {"snaktype": "value", "property": "P1889", "hash": "fa4d7f858821ba2ba28bcf96f8f4dde9d33a40b4", "datavalue": {"value": {"entity-type": "item", "numeric-id": 47574, "id": "Q47574"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q2198779$91424fa3-4e4a-14a2-c39d-d205f398ff4d", "rank": "normal"}], "P443": [{"mainsnak": {"snaktype": "value", "property": "P443", "hash": "b2586b4e1a97d29c3314fc91ee3bb7e88aa7a646", "datavalue": {"value": "De-Einheit.ogg", "type": "string"}, "datatype": "commonsMedia"}, "type": "statement", "qualifiers": {"P407": [{"snaktype": "value", "property": "P407", "hash": "46bfd327b830f66f7061ea92d1be430c135fa91f", "datavalue": {"value": {"entity-type": "item", "numeric-id": 188, "id": "Q188"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}], "P898": [{"snaktype": "value", "property": "P898", "hash": "91460a55f5c5724a15a541241350fcb1667b9a2c", "datavalue": {"value": "\u02c8a\u026a\u032fnha\u026a\u032ft", "type": "string"}, "datatype": "string"}]}, "qualifiers-order": ["P407", "P898"], "id": "Q2198779$7d271015-475a-371d-e8b9-c6a8c5f533a0", "rank": "normal"}]}}}, "success": 1}
