+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q23958852.json
@@ -1,0 +1,1 @@
+{"entities": {"Q23958852": {"pageid": 25953807, "ns": 0, "title": "Q23958852", "lastrevid": 2321950972, "modified": "2025-03-09T02:17:29Z", "type": "item", "id": "Q23958852", "labels": {"en": {"language": "en", "value": "variable-order class"}}, "descriptions": {"en": {"language": "en", "value": "abstract entity representing classes whose instances may be classes of different orders or even non-classes"}}, "aliases": {}, "claims": {"P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "33c0fd1431aa0fb276f87bb51b4a956d9440bb15", "datavalue": {"value": {"entity-type": "item", "numeric-id": 16889133, "id": "Q16889133"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q23958852$e3c762f7-4ab3-8a2b-7bf0-f411b7f135fc", "rank": "normal"}], "P1343": [{"mainsnak": {"snaktype": "value", "property": "P1343", "hash": "5d5a0071814b32cd11925e928f290e684abc722d", "datavalue": {"value": {"entity-type": "item", "numeric-id": 23959013, "id": "Q23959013"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q23958852$6701f85e-4030-5ff3-6fd4-268a21e63c7c", "rank": "normal"}], "P1709": [{"mainsnak": {"snaktype": "value", "property": "P1709", "hash": "fe4c6429300867e659f1c542d39cbb636ab6a971", "datavalue": {"value": "http://sw.opencyc.org/2012/05/10/concept/en/VariedOrderCollection", "type": "string"}, "datatype": "url"}, "type": "statement", "id": "Q23958852$59dbc8de-4abf-a010-f5fb-369dc50b94c4", "rank": "normal"}], "P2559": [{"mainsnak": {"snaktype": "value", "property": "P2559", "hash": "581e8db072b0708d411f2e05971a0668d928f2e6", "datavalue": {"value": {"text": "A variable-order class in Wikidata, e.g., class (Q16889133) or variable-order class (Q23958852), is a class that is not a fixed-order class, e.g., one whose instances can include both first-order classes and second-order classes. There are only a few natural variable-order classes.  Before stating that a class is an instance of (P31) variable-order class (Q23958852) check that it is actually a variable-order class, e.g., that it can contain both classes and non-classes.", "language": "en"}, "type": "monolingualtext"}, "datatype": "monolingualtext"}, "type": "statement", "id": "Q23958852$501e18d5-41b9-b08c-776e-a476922b689a", "rank": "normal"}], "P31": [{"mainsnak": {"snaktype": "value", "property": "P31", "hash": "9d31d99ce62f7f1424d6d2be409f25a62397445f", "datavalue": {"value": {"entity-type": "item", "numeric-id": 23958852, "id": "Q23958852"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q23958852$4101a465-4164-37c6-f564-07fa69f64576", "rank": "normal", "references": [{"hash": "be24ec307509c95ce446494e789ac8f2c69fd73c", "snaks": {"P248": [{"snaktype": "value", "property": "P248", "hash": "9ee8cec39cf8b05a4c3c12d323e5dadcccb87782", "datavalue": {"value": {"entity-type": "item", "numeric-id": 23959013, "id": "Q23959013"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "snaks-order": ["P248"]}]}], "P5869": [{"mainsnak": {"snaktype": "value", "property": "P5869", "hash": "d2f4941175039aef5c28a19ad773481061d431b0", "datavalue": {"value": {"entity-type": "item", "numeric-id": 35120, "id": "Q35120"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P11527": [{"snaktype": "value", "property": "P11527", "hash": "03333cd9754e182cbd8cf05cf461c34d428ffcbe", "datavalue": {"value": {"entity-type": "property", "numeric-id": 31, "id": "P31"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}]}, "qualifiers-order": ["P11527"], "id": "Q23958852$2d7f8fd0-4241-f4c5-335e-d9a4e2ffb22c", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P5869", "hash": "42bef09269037c1949682d6de8ce6ea726f8014f", "datavalue": {"value": {"entity-type": "item", "numeric-id": 112234259, "id": "Q112234259"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P11527": [{"snaktype": "value", "property": "P11527", "hash": "03333cd9754e182cbd8cf05cf461c34d428ffcbe", "datavalue": {"value": {"entity-type": "property", "numeric-id": 31, "id": "P31"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}]}, "qualifiers-order": ["P11527"], "id": "Q23958852$206cdbe8-4030-ec42-9fa1-7ad9d4c8a4c0", "rank": "normal"}], "P6104": [{"mainsnak": {"snaktype": "value", "property": "P6104", "hash": "2bcf963484ee556d691849e71ea3ae951738fc36", "datavalue": {"value": {"entity-type": "item", "numeric-id": 60005226, "id": "Q60005226"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q23958852$78B86424-6E0A-469A-8DF1-0BD0EB34F8EB", "rank": "normal"}]}}}, "success": 1}
