+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q23959932.json
@@ -1,0 +1,1 @@
+{"entities": {"Q23959932": {"pageid": 25954876, "ns": 0, "title": "Q23959932", "lastrevid": 2298845101, "modified": "2025-01-16T18:23:49Z", "type": "item", "id": "Q23959932", "labels": {"en": {"language": "en", "value": "fixed-order class"}}, "descriptions": {"en": {"language": "en", "value": "class whose instances are all either individuals or same-fixed-order classes"}}, "aliases": {}, "claims": {"P1709": [{"mainsnak": {"snaktype": "value", "property": "P1709", "hash": "d6e2b70d8782a2318760aabcb0e08aedeb39456a", "datavalue": {"value": "http://sw.opencyc.org/2012/05/10/concept/en/FixedOrderCollection", "type": "string"}, "datatype": "url"}, "type": "statement", "id": "Q23959932$20798852-46ca-4657-cdaf-09ae9c949130", "rank": "normal"}], "P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "33c0fd1431aa0fb276f87bb51b4a956d9440bb15", "datavalue": {"value": {"entity-type": "item", "numeric-id": 16889133, "id": "Q16889133"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q23959932$207f587d-49ef-27cf-717b-1b8734037a59", "rank": "normal"}], "P2738": [{"mainsnak": {"snaktype": "value", "property": "P2738", "hash": "4fa7d4837acece3e3ee87aa78037ef4527696ea4", "datavalue": {"value": {"entity-type": "item", "numeric-id": 23766486, "id": "Q23766486"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P11260": [{"snaktype": "value", "property": "P11260", "hash": "7134c0d3a164bcbef63a82a231ad07eae94f5938", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24017414, "id": "Q24017414"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, {"snaktype": "value", "property": "P11260", "hash": "848b71d9ecce40c52a7bed662829f2ed0c5f0da1", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24017465, "id": "Q24017465"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, {"snaktype": "value", "property": "P11260", "hash": "2e2e51efba85c8d99492f080a6f2d226cff8fcd4", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24027474, "id": "Q24027474"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, {"snaktype": "value", "property": "P11260", "hash": "235e0edc6853e17b89e871edae8b25fa0e494645", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24027515, "id": "Q24027515"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, {"snaktype": "value", "property": "P11260", "hash": "5e7ad69b8300d9139b450e3d6e171f2fa8381035", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24027526, "id": "Q24027526"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, {"snaktype": "value", "property": "P11260", "hash": "0a4157dee08ad8168b485594e92cfe4a6e5b5750", "datavalue": {"value": {"entity-type": "item", "numeric-id": 104086571, "id": "Q104086571"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P11260"], "id": "Q23959932$3e44b7e8-46d8-46ad-644b-5fbb9c5ef5da", "rank": "normal"}], "P5008": [{"mainsnak": {"snaktype": "value", "property": "P5008", "hash": "25626b3f3cf752a34ec1e4705ba2288013a84078", "datavalue": {"value": {"entity-type": "item", "numeric-id": 68235346, "id": "Q68235346"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q23959932$6E12562B-9F75-4555-902B-13548B121F14", "rank": "normal"}], "P31": [{"mainsnak": {"snaktype": "value", "property": "P31", "hash": "9d31d99ce62f7f1424d6d2be409f25a62397445f", "datavalue": {"value": {"entity-type": "item", "numeric-id": 23958852, "id": "Q23958852"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q23959932$ea0166a7-4f7d-7442-a52d-fc581ce5b585", "rank": "normal"}], "P2559": [{"mainsnak": {"snaktype": "value", "property": "P2559", "hash": "59196f8abe8f799944e3c21bf646bf53b61d5b77", "datavalue": {"value": {"text": "A fixed-order class in Wikidata, e.g.,  watercraft (Q1229765),  watercraft type (Q16335899), or second-order class (Q24017414), is a class whose instances are all of the same order, e.g., all not classes (a  first-order class (Q104086571) like human (Q5)) or all classes whose instances are not classes (a second-order class (Q24017414) like ship type (Q2235308)).No class should be explicitly an instance of (P31) fixed-order class (Q23959932); instead use one of first-order class (Q104086571), second-order class (Q24017414), third-order class (Q24017465), or very rarely one of the other subclasses of fixed-order class (Q23959932).", "language": "en"}, "type": "monolingualtext"}, "datatype": "monolingualtext"}, "type": "statement", "id": "Q23959932$e9b2f65d-499f-1dc4-1213-7f401f297a3a", "rank": "normal"}], "P6104": [{"mainsnak": {"snaktype": "value", "property": "P6104", "hash": "2bcf963484ee556d691849e71ea3ae951738fc36", "datavalue": {"value": {"entity-type": "item", "numeric-id": 60005226, "id": "Q60005226"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q23959932$BF1B19FA-DEC0-490A-B512-4690D3E2555A", "rank": "normal"}]}}}, "success": 1}
