+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q24017465.json
@@ -1,0 +1,1 @@
+{"entities": {"Q24017465": {"pageid": 26010552, "ns": 0, "title": "Q24017465", "lastrevid": 2309068700, "modified": "2025-02-09T11:55:57Z", "type": "item", "id": "Q24017465", "labels": {"en": {"language": "en", "value": "third-order class"}}, "descriptions": {"en": {"language": "en", "value": "metaclass containing as instances all classes of classes of individuals"}}, "aliases": {"en": [{"language": "en", "value": "3rd-order class"}, {"language": "en", "value": "class of the third order"}, {"language": "en", "value": "class of the 3rd order"}, {"language": "en", "value": "3rd-order metaclass"}, {"language": "en", "value": "metaclass of the third order"}, {"language": "en", "value": "metaclass of the 3rd order"}, {"language": "en", "value": "third-order metaclass"}, {"language": "en", "value": "metametaclass"}]}, "claims": {"P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "c1a4184b5aab81dd72dd84a384a7ab3522eaa365", "datavalue": {"value": {"entity-type": "item", "numeric-id": 23959932, "id": "Q23959932"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P1545": [{"snaktype": "value", "property": "P1545", "hash": "0e979f28bf306fefdcd352b4eb8dee5da2153a6d", "datavalue": {"value": "3", "type": "string"}, "datatype": "string"}], "P155": [{"snaktype": "value", "property": "P155", "hash": "e47e61d36e9ba2068a5e81f72971d7dde9b93ba6", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24017414, "id": "Q24017414"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}], "P156": [{"snaktype": "value", "property": "P156", "hash": "25dd6eab4e158a8f3af839f5b782b4b1b3a73795", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24027474, "id": "Q24027474"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P1545", "P155", "P156"], "id": "Q24017465$778a6655-42fe-74a7-53ac-0c575826f884", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P279", "hash": "cf9c07e5b85d515758fb633dd79aea6782279bcd", "datavalue": {"value": {"entity-type": "item", "numeric-id": 19478619, "id": "Q19478619"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q24017465$146b6712-497c-ba84-4105-c9dd79b3d24e", "rank": "normal"}], "P1709": [{"mainsnak": {"snaktype": "value", "property": "P1709", "hash": "bd61c48877df0092ce23e6a5c1cd6a665e5e7071", "datavalue": {"value": "http://sw.opencyc.org/concept/Mx4rHQdVmB_TEdaAAABQ2rksLw", "type": "string"}, "datatype": "url"}, "type": "statement", "id": "Q24017465$0c8ebb64-4af4-b2b2-0c0e-3533190b24b5", "rank": "normal"}], "P31": [{"mainsnak": {"snaktype": "value", "property": "P31", "hash": "1004694dfba0d4f3d4f4a59e59933734a4b3452c", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24027474, "id": "Q24027474"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q24017465$0afe501e-4efd-c6a9-c2d7-487b28922dfa", "rank": "normal"}], "P1963": [{"mainsnak": {"snaktype": "value", "property": "P1963", "hash": "64314ae3d2ba8aaa19b7d900e52d4f41a7e6ddf1", "datavalue": {"value": {"entity-type": "property", "numeric-id": 8225, "id": "P8225"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}, "type": "statement", "id": "Q24017465$f19ac53f-4505-eb29-5e3f-923c4cdf9508", "rank": "normal"}], "P11012": [{"mainsnak": {"snaktype": "value", "property": "P11012", "hash": "6b52a6f139d9e29248564e7c28d0914966533223", "datavalue": {"value": "Q3996892", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q24017465$F97F10C3-33E3-46BA-81D6-7B221C5D8AAF", "rank": "normal"}], "P2559": [{"mainsnak": {"snaktype": "value", "property": "P2559", "hash": "d5cb34bfeee59d1fb795fdbaa00f775e95ff6a6b", "datavalue": {"value": {"text": "A third-order class in Wikidata, e.g., vehicle model series (Q29048319) or second-order class (Q24017414), is a class all of whose instances are implicitly or explicitly instances of (P31) second-order class (Q24017414).  A class is implicitly a third-order classes if it is a subclass of (P279) another class that is a third-order class.  These implicit instance of (P31) statements are not all observable in Wikidata.There are very few natural third-order classes.  Before stating that a class is instance of (P31) third-order class (Q24017465) check that it is actually a third-order class.", "language": "en"}, "type": "monolingualtext"}, "datatype": "monolingualtext"}, "type": "statement", "id": "Q24017465$84a7ac19-471c-f544-3bf9-0ce7d8ba6707", "rank": "normal"}], "P8225": [{"mainsnak": {"snaktype": "value", "property": "P8225", "hash": "ccd23585ad3f1ceee43bb32e0912f8cd1d79749a", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24017414, "id": "Q24017414"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q24017465$0ae0bfde-4110-5dd3-5e58-c08e27c77de2", "rank": "normal"}], "P6104": [{"mainsnak": {"snaktype": "value", "property": "P6104", "hash": "2bcf963484ee556d691849e71ea3ae951738fc36", "datavalue": {"value": {"entity-type": "item", "numeric-id": 60005226, "id": "Q60005226"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q24017465$663D7506-21B9-4E6D-8C7C-A8BCE0741A0C", "rank": "normal"}]}}}, "success": 1}
