+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q26401003.json
@@ -1,0 +1,1 @@
+{"entities": {"Q26401003": {"pageid": 28351878, "ns": 0, "title": "Q26401003", "lastrevid": 2297445733, "modified": "2025-01-12T19:03:28Z", "type": "item", "id": "Q26401003", "labels": {"en": {"language": "en", "value": "individual animal"}}, "descriptions": {"en": {"language": "en", "value": "singular named exemplar of an animal (e.g., the gorilla named Koko; the cat named Socks)"}}, "aliases": {"en": [{"language": "en", "value": "famous animal"}, {"language": "en", "value": "animal individual"}, {"language": "en", "value": "instance of individual animal"}, {"language": "en", "value": "instance of animal individual"}, {"language": "en", "value": "unique animal"}]}, "claims": {"P910": [{"mainsnak": {"snaktype": "value", "property": "P910", "hash": "74fc44016256b26d56131251e742472092be35be", "datavalue": {"value": {"entity-type": "item", "numeric-id": 7157842, "id": "Q7157842"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q26401003$f83b8e66-469d-8cc0-76e8-252fe743dd8d", "rank": "normal"}], "P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "c0f9d3b7ec2d5b2cb10db8be2908730598b30462", "datavalue": {"value": {"entity-type": "item", "numeric-id": 10855152, "id": "Q10855152"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q26401003$ea401810-4007-b114-e4e0-d2fff86c1d32", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P279", "hash": "7d74e853742d8452a14e9dab0438f67c22cf62b1", "datavalue": {"value": {"entity-type": "item", "numeric-id": 729, "id": "Q729"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q26401003$cb619aa2-4852-18de-2e78-b3fa8e888ffd", "rank": "normal"}], "P373": [{"mainsnak": {"snaktype": "value", "property": "P373", "hash": "6c2fa489a45016bc0313b3f827151976f938c42a", "datavalue": {"value": "Famous animals", "type": "string"}, "datatype": "string"}, "type": "statement", "id": "Q26401003$dfe70a45-4cff-447f-25af-8560b89c2d9b", "rank": "normal"}], "P1424": [{"mainsnak": {"snaktype": "value", "property": "P1424", "hash": "e0e703802e39010874247b27a50f3668b1f1d78e", "datavalue": {"value": {"entity-type": "item", "numeric-id": 12681112, "id": "Q12681112"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q26401003$50d6b1ca-438f-32a4-524b-070d677e2f0f", "rank": "normal"}], "P1963": [{"mainsnak": {"snaktype": "value", "property": "P1963", "hash": "cf74ee35f9a696c4345504c2121822533335355e", "datavalue": {"value": {"entity-type": "property", "numeric-id": 10241, "id": "P10241"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}, "type": "statement", "id": "Q26401003$98ca05fd-4687-c6e8-497d-473f46cb7757", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P1963", "hash": "be6d5b6e6a3a8c6e7ad2166fbb08237efaf27b52", "datavalue": {"value": {"entity-type": "property", "numeric-id": 21, "id": "P21"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}, "type": "statement", "id": "Q26401003$80b413ab-4465-9760-89d1-3d7f3a03799f", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P1963", "hash": "c58af33a98a34cdf63f21b804fb5b50503e461bb", "datavalue": {"value": {"entity-type": "property", "numeric-id": 569, "id": "P569"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}, "type": "statement", "id": "Q26401003$6f740f12-4055-fbf0-811b-641a02e1aed8", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P1963", "hash": "e9b6f0f63af10cf5f91a1b61c5fd5fc8977a5b26", "datavalue": {"value": {"entity-type": "property", "numeric-id": 19, "id": "P19"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}, "type": "statement", "id": "Q26401003$cbf87bf8-49d7-8e5b-0978-b16465c5f7f2", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P1963", "hash": "058a6ebaad6ae6a4a7b616fbd88f415015a523ae", "datavalue": {"value": {"entity-type": "property", "numeric-id": 570, "id": "P570"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}, "type": "statement", "id": "Q26401003$4e5cfde6-4eae-6e49-e1c7-d631f33716b0", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P1963", "hash": "f925dcfe282c872d92525ef7fc8e25630cb45318", "datavalue": {"value": {"entity-type": "property", "numeric-id": 20, "id": "P20"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}, "type": "statement", "id": "Q26401003$9fb08f8c-41e0-5288-4524-f54ebb228dce", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P1963", "hash": "6e69ad37e0a773858bb44fd1a1cf74cb7bc780fe", "datavalue": {"value": {"entity-type": "property", "numeric-id": 138, "id": "P138"}, "type": "wikibase-entityid"}, "datatype": "wikibase-property"}, "type": "statement", "id": "Q26401003$9c76926b-416b-b53a-b24f-ac9b3da47ff3", "rank": "normal"}]}}}, "success": 1}
