+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q4330518.json
@@ -1,0 +1,1 @@
+{"entities": {"Q4330518": {"pageid": 4138790, "ns": 0, "title": "Q4330518", "lastrevid": 2277380245, "modified": "2024-11-20T07:36:48Z", "type": "item", "id": "Q4330518", "labels": {"en": {"language": "en", "value": "object of study"}}, "descriptions": {"en": {"language": "en", "value": "something that undergoes testing or observation in a scientific experiment or in a scientific study"}}, "aliases": {"en": [{"language": "en", "value": "object of research"}, {"language": "en", "value": "research object"}, {"language": "en", "value": "study object"}]}, "claims": {"P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "74e8713f3f62eacc5779f5f6fe44ca33b414cdb8", "datavalue": {"value": {"entity-type": "item", "numeric-id": 26256810, "id": "Q26256810"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q4330518$838ff3dc-4d9b-2cb9-dc99-c60f97dadce1", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P279", "hash": "7999a38f46ebb0d75beef6668b45df097ddbde3b", "datavalue": {"value": {"entity-type": "item", "numeric-id": 123177423, "id": "Q123177423"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P2210": [{"snaktype": "value", "property": "P2210", "hash": "fd3334934d75df7c92b4990b21c16d808fe8dfa2", "datavalue": {"value": {"entity-type": "item", "numeric-id": 42240, "id": "Q42240"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P2210"], "id": "Q4330518$a6c29325-4fa4-6a8f-06a2-866e69a80e2f", "rank": "normal"}], "P361": [{"mainsnak": {"snaktype": "value", "property": "P361", "hash": "66482ad85f642b9d419dc215ae874902384fd1e3", "datavalue": {"value": {"entity-type": "item", "numeric-id": 42240, "id": "Q42240"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q4330518$b897a894-4c22-81bc-0949-4c586ace4945", "rank": "normal"}], "P2671": [{"mainsnak": {"snaktype": "value", "property": "P2671", "hash": "40679ac013ae1345a2d3339c644049eb991cc61f", "datavalue": {"value": "/g/121m_h9t", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q4330518$A21D6E5A-49AF-44BF-BD04-B8BD783BA993", "rank": "normal"}], "P2347": [{"mainsnak": {"snaktype": "value", "property": "P2347", "hash": "992fba781378c8c3b1d5b84e155178ae0494d219", "datavalue": {"value": "23194", "type": "string"}, "datatype": "external-id"}, "type": "statement", "qualifiers": {"P1810": [{"snaktype": "value", "property": "P1810", "hash": "1b8add0c9eb87b22f605734a162901291e40b942", "datavalue": {"value": "tutkimuskohteet", "type": "string"}, "datatype": "string"}, {"snaktype": "value", "property": "P1810", "hash": "42e2f77b25ae6acbfe596c156d318ac04e863979", "datavalue": {"value": "studieobjekt", "type": "string"}, "datatype": "string"}, {"snaktype": "value", "property": "P1810", "hash": "c4fb0923da61cc0f5682e5cb7780a583bc43a6be", "datavalue": {"value": "research subjects", "type": "string"}, "datatype": "string"}]}, "qualifiers-order": ["P1810"], "id": "Q4330518$9ADF2173-0ADA-4F2C-9EC0-799D56937BA5", "rank": "normal", "references": [{"hash": "d864245f0775b15ea147d7f55e46111bcbaf89d4", "snaks": {"P248": [{"snaktype": "value", "property": "P248", "hash": "7133f11674741f52cadaae6029068fad9cbb52e3", "datavalue": {"value": {"entity-type": "item", "numeric-id": 89345680, "id": "Q89345680"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}], "P813": [{"snaktype": "value", "property": "P813", "hash": "d14ae00ee696ad55b3897a3fc9cf64bce5c78410", "datavalue": {"value": {"time": "+2022-07-22T00:00:00Z", "timezone": 0, "before": 0, "after": 0, "precision": 11, "calendarmodel": "http://www.wikidata.org/entity/Q1985727"}, "type": "time"}, "datatype": "time"}]}, "snaks-order": ["P248", "P813"]}]}], "P31": [{"mainsnak": {"snaktype": "value", "property": "P31", "hash": "9f199485b3027d9f715fa2b0f736d7d4ace3295a", "datavalue": {"value": {"entity-type": "item", "numeric-id": 96251598, "id": "Q96251598"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q4330518$aeb459d1-49d4-dbba-4cd5-983fdf128b2e", "rank": "normal"}], "P527": [{"mainsnak": {"snaktype": "value", "property": "P527", "hash": "7e2b9d7c586df94f7bf80777ebcd54c83437ab51", "datavalue": {"value": {"entity-type": "item", "numeric-id": 309, "id": "Q309"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q4330518$3b4a4862-48b4-b64d-0666-d5a4b1cde867", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P527", "hash": "6e3fc30f3a2414699398d95b6eeee11167bf6be5", "datavalue": {"value": {"entity-type": "item", "numeric-id": 7887142, "id": "Q7887142"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q4330518$0474fda0-46c7-c9ef-6336-8d9e7307c140", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P527", "hash": "e688fb06f3bcdfcf3d62cd097e095382ef068b18", "datavalue": {"value": {"entity-type": "item", "numeric-id": 213907, "id": "Q213907"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q4330518$d187e430-4066-8853-febd-36159d4678d5", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P527", "hash": "2f670412f9523c7519befe5775a54ce9a4c46545", "datavalue": {"value": {"entity-type": "item", "numeric-id": 55596814, "id": "Q55596814"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q4330518$9c074846-43d1-ccfd-4e3e-43af1e0d1555", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P527", "hash": "04c4eeaff0b4c498a49abdccac2b518fdcbd5f1c", "datavalue": {"value": {"entity-type": "item", "numeric-id": 2148398, "id": "Q2148398"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q4330518$28e2bdb7-4657-390c-feca-c6bcd79606ba", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P527", "hash": "0a7fbdca7c9e3df421eb72d9504887770c4c0b07", "datavalue": {"value": {"entity-type": "item", "numeric-id": 4204239, "id": "Q4204239"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q4330518$83e803eb-4a5d-211e-5e67-b52f1c6be8b2", "rank": "normal"}]}}}, "success": 1}
