+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q53617489.json
@@ -1,0 +1,1 @@
+{"entities": {"Q53617489": {"pageid": 54123447, "ns": 0, "title": "Q53617489", "lastrevid": 2324681226, "modified": "2025-03-14T18:47:18Z", "type": "item", "id": "Q53617489", "labels": {"en": {"language": "en", "value": "independent continuant"}}, "descriptions": {"en": {"language": "en", "value": "BFO class"}}, "aliases": {}, "claims": {"P2888": [{"mainsnak": {"snaktype": "value", "property": "P2888", "hash": "4f62d8d3cbfd17366c478d314c33a6b346765e8f", "datavalue": {"value": "http://purl.obolibrary.org/obo/BFO_0000004", "type": "string"}, "datatype": "url"}, "type": "statement", "id": "Q53617489$5cf825f5-47e1-3d73-b101-87865253f0d4", "rank": "normal"}], "P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "5be7ebd8b2dd1ade3648bf7dfd0843e8a3959f00", "datavalue": {"value": {"entity-type": "item", "numeric-id": 103940464, "id": "Q103940464"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q53617489$468bc1d1-4da8-35f9-4f45-de5986e9f6c7", "rank": "normal"}], "P31": [{"mainsnak": {"snaktype": "value", "property": "P31", "hash": "1c5e8fa6972fb543163f82f52d2e8b3211b7bd91", "datavalue": {"value": {"entity-type": "item", "numeric-id": 124711104, "id": "Q124711104"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q53617489$1a053a2e-4c3b-27ed-350b-4187a142d101", "rank": "normal"}], "P2738": [{"mainsnak": {"snaktype": "value", "property": "P2738", "hash": "4fa7d4837acece3e3ee87aa78037ef4527696ea4", "datavalue": {"value": {"entity-type": "item", "numeric-id": 23766486, "id": "Q23766486"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P11260": [{"snaktype": "value", "property": "P11260", "hash": "ead782dfa538b9d0177585e2efd251e8ec98e49a", "datavalue": {"value": {"entity-type": "item", "numeric-id": 53617407, "id": "Q53617407"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, {"snaktype": "value", "property": "P11260", "hash": "f8fdd5fa6789bc3d110f6d2a124cbfdbbc09e1f0", "datavalue": {"value": {"entity-type": "item", "numeric-id": 124711467, "id": "Q124711467"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P11260"], "id": "Q53617489$69124720-4aaa-1d80-2754-ab2fed22c43c", "rank": "normal"}], "P12602": [{"mainsnak": {"snaktype": "value", "property": "P12602", "hash": "19f6d01e0d24dfa6e6aa320b5dc0c39c0bad123e", "datavalue": {"value": "0000004", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q53617489$e0270358-49f0-1d79-cbe8-a56e7fcbeef5", "rank": "normal"}]}}}, "success": 1}
