+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q66394244.json
@@ -1,0 +1,1 @@
+{"entities": {"Q66394244": {"pageid": 65995274, "ns": 0, "title": "Q66394244", "lastrevid": 2296925219, "modified": "2025-01-11T09:08:36Z", "type": "item", "id": "Q66394244", "labels": {"en": {"language": "en", "value": "physical anatomical entity"}}, "descriptions": {"en": {"language": "en", "value": "unit of member of a species's structural organization"}}, "aliases": {"en": [{"language": "en", "value": "anatomical object"}]}, "claims": {"P1402": [{"mainsnak": {"snaktype": "value", "property": "P1402", "hash": "a875d247d979fb4139d2defbbedc6c5ab90ee90c", "datavalue": {"value": "61775", "type": "string"}, "datatype": "external-id"}, "type": "statement", "qualifiers": {"P1810": [{"snaktype": "value", "property": "P1810", "hash": "7427cc5dcd44c97fec5a8d756afb23ad8338ce97", "datavalue": {"value": "Physical anatomical entity", "type": "string"}, "datatype": "string"}]}, "qualifiers-order": ["P1810"], "id": "Q66394244$18747517-4221-c1b8-3572-fa27ea460722", "rank": "normal", "references": [{"hash": "3deb5336f981936e88d7ecccdfbb3519c4e7e149", "snaks": {"P248": [{"snaktype": "value", "property": "P248", "hash": "1b5fd2ff766b14dac1963a5e8745c7bdf17e9323", "datavalue": {"value": {"entity-type": "item", "numeric-id": 1406710, "id": "Q1406710"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}], "P813": [{"snaktype": "value", "property": "P813", "hash": "d20c80df1d123bd85fffeee7c590024140499da9", "datavalue": {"value": {"time": "+2019-08-01T00:00:00Z", "timezone": 0, "before": 0, "after": 0, "precision": 11, "calendarmodel": "http://www.wikidata.org/entity/Q1985727"}, "type": "time"}, "datatype": "time"}]}, "snaks-order": ["P248", "P813"]}]}], "P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "b2fb48849d515c1c268da981b14decfab0fc3b6a", "datavalue": {"value": {"entity-type": "item", "numeric-id": 27043950, "id": "Q27043950"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q66394244$525c3db0-4f59-5714-6068-216d0deee23f", "rank": "normal", "references": [{"hash": "c519e62eba6ffeb2434877a2c9889f588fa7b21e", "snaks": {"P248": [{"snaktype": "value", "property": "P248", "hash": "1b5fd2ff766b14dac1963a5e8745c7bdf17e9323", "datavalue": {"value": {"entity-type": "item", "numeric-id": 1406710, "id": "Q1406710"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "snaks-order": ["P248"]}, {"hash": "afd93b03d4c57dd8c14d4c66f7bea61701ba5985", "snaks": {"P1402": [{"snaktype": "value", "property": "P1402", "hash": "a875d247d979fb4139d2defbbedc6c5ab90ee90c", "datavalue": {"value": "61775", "type": "string"}, "datatype": "external-id"}], "P248": [{"snaktype": "value", "property": "P248", "hash": "1b5fd2ff766b14dac1963a5e8745c7bdf17e9323", "datavalue": {"value": {"entity-type": "item", "numeric-id": 1406710, "id": "Q1406710"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}], "P813": [{"snaktype": "value", "property": "P813", "hash": "d20c80df1d123bd85fffeee7c590024140499da9", "datavalue": {"value": {"time": "+2019-08-01T00:00:00Z", "timezone": 0, "before": 0, "after": 0, "precision": 11, "calendarmodel": "http://www.wikidata.org/entity/Q1985727"}, "type": "time"}, "datatype": "time"}]}, "snaks-order": ["P1402", "P248", "P813"]}]}, {"mainsnak": {"snaktype": "value", "property": "P279", "hash": "21550c6e2ea515c81623f6987e296de66ce54d50", "datavalue": {"value": {"entity-type": "item", "numeric-id": 4406616, "id": "Q4406616"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q66394244$56b4a260-470a-9312-38ef-08f01238e8d4", "rank": "normal"}], "P31": [{"mainsnak": {"snaktype": "value", "property": "P31", "hash": "e7b1f703ea74afb2601e84cd39a91a0c53708861", "datavalue": {"value": {"entity-type": "item", "numeric-id": 104007623, "id": "Q104007623"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q66394244$5828dec3-4777-55d7-e34b-d133f4edba9e", "rank": "normal"}], "P461": [{"mainsnak": {"snaktype": "value", "property": "P461", "hash": "745f11cde9d038db0ead39aec7db415ac7cc612a", "datavalue": {"value": {"entity-type": "item", "numeric-id": 66562954, "id": "Q66562954"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q66394244$8d1e9ac1-452a-5d52-68a6-cbd634b149fd", "rank": "normal"}], "P460": [{"mainsnak": {"snaktype": "value", "property": "P460", "hash": "a73e22755492d568964aa85239fd0254a72c67e7", "datavalue": {"value": {"entity-type": "item", "numeric-id": 27043948, "id": "Q27043948"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q66394244$2d768a7f-4d35-87a7-2df9-b2af68bd92b9", "rank": "normal"}]}}}, "success": 1}
