+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q72638.json
@@ -1,0 +1,1 @@
+{"entities": {"Q72638": {"pageid": 75326, "ns": 0, "title": "Q72638", "lastrevid": 2161741408, "modified": "2024-05-22T15:55:57Z", "type": "item", "id": "Q72638", "labels": {"en": {"language": "en", "value": "consumer"}}, "descriptions": {"en": {"language": "en", "value": "organism in an ecological food chain that receives energy by consuming other organisms"}}, "aliases": {}, "claims": {"P646": [{"mainsnak": {"snaktype": "value", "property": "P646", "hash": "a0d09d29320cd93fa65d9d7de38ae0344ed6df3c", "datavalue": {"value": "/m/07kd152", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q72638$8E61B10B-1F2B-453A-B098-FCB8FAC12C8E", "rank": "normal", "references": [{"hash": "2b00cb481cddcac7623114367489b5c194901c4a", "snaks": {"P248": [{"snaktype": "value", "property": "P248", "hash": "a94b740202b097dd33355e0e6c00e54b9395e5e0", "datavalue": {"value": {"entity-type": "item", "numeric-id": 15241312, "id": "Q15241312"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}], "P577": [{"snaktype": "value", "property": "P577", "hash": "fde79ecb015112d2f29229ccc1ec514ed3e71fa2", "datavalue": {"value": {"time": "+2013-10-28T00:00:00Z", "timezone": 0, "before": 0, "after": 0, "precision": 11, "calendarmodel": "http://www.wikidata.org/entity/Q1985727"}, "type": "time"}, "datatype": "time"}]}, "snaks-order": ["P248", "P577"]}]}], "P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "cdf462c5ea1c5d17a0e32400558c8059df2db8ba", "datavalue": {"value": {"entity-type": "item", "numeric-id": 159344, "id": "Q159344"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q72638$4a797aa8-4e5b-87f4-3c62-8b789c214b16", "rank": "normal"}], "P2924": [{"mainsnak": {"snaktype": "value", "property": "P2924", "hash": "68f0c72c7e6c9744bed35dbc4604162d6efd3438", "datavalue": {"value": "2092896", "type": "string"}, "datatype": "external-id"}, "type": "statement", "qualifiers": {"P1810": [{"snaktype": "value", "property": "P1810", "hash": "f28fc0b04c50102781e7db6f9153aac52fd8739c", "datavalue": {"value": "\u041a\u041e\u041d\u0421\u0423\u041c\u0415\u041d\u0422\u042b", "type": "string"}, "datatype": "string"}]}, "qualifiers-order": ["P1810"], "id": "Q72638$BB8F03B7-1E96-4984-9BCC-D7D875190164", "rank": "normal"}], "P6366": [{"mainsnak": {"snaktype": "value", "property": "P6366", "hash": "91cf4bd42faec6bfd9351e9be497b0e11dab0aef", "datavalue": {"value": "28864654", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q72638$4827DD41-EF58-43BF-A1F5-8CF838330DFA", "rank": "normal"}], "P7033": [{"mainsnak": {"snaktype": "value", "property": "P7033", "hash": "2388e79d513e854d97746d277dd5a6309f6d9b2c", "datavalue": {"value": "scot/4352", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q72638$0F5FEE77-696A-4C20-9C58-13E5FC352B41", "rank": "normal"}], "P1889": [{"mainsnak": {"snaktype": "value", "property": "P1889", "hash": "9a15955f90f7fc4fafa23384820b5d3c9b5c01a7", "datavalue": {"value": {"entity-type": "item", "numeric-id": 342139, "id": "Q342139"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P1013": [{"snaktype": "value", "property": "P1013", "hash": "3ee85f87aa07bea0f614c2220e1ccdf8d25de085", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24005632, "id": "Q24005632"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P1013"], "id": "Q72638$386B111C-CDC0-4EDB-A78C-946DF960F380", "rank": "normal", "references": [{"hash": "ceca8f2917accbb8fe82c7e04d65aed35128b629", "snaks": {"P143": [{"snaktype": "value", "property": "P143", "hash": "39224a9c2e8ce5424defbd16603d25771956c7fc", "datavalue": {"value": {"entity-type": "item", "numeric-id": 1551807, "id": "Q1551807"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}], "P4656": [{"snaktype": "value", "property": "P4656", "hash": "acd4c5108ce63b8a7644f922192c7ce025793eef", "datavalue": {"value": "https://pl.wikipedia.org/w/index.php?title=Konsument_(biologia)&oldid=59347940", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P143", "P4656"]}]}], "P8313": [{"mainsnak": {"snaktype": "value", "property": "P8313", "hash": "1fe52d11eaa45c8c5363f8664e3e0f8a33d17dd5", "datavalue": {"value": "konsument", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q72638$0FBDA308-2059-4B38-99F4-8B2C16849917", "rank": "normal"}], "P4342": [{"mainsnak": {"snaktype": "value", "property": "P4342", "hash": "dc22de7ae0b0b039f6c649251a693395a8e83a18", "datavalue": {"value": "konsument_-_organisme", "type": "string"}, "datatype": "external-id"}, "type": "statement", "qualifiers": {"P4390": [{"snaktype": "value", "property": "P4390", "hash": "1a4df62914ea9afca349bd5fb5d8efd5832d83fa", "datavalue": {"value": {"entity-type": "item", "numeric-id": 39893449, "id": "Q39893449"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}], "P1810": [{"snaktype": "value", "property": "P1810", "hash": "78df4c3a94c5cc0a81c1ffe3c288a6d4da05681c", "datavalue": {"value": "konsument (organisme)", "type": "string"}, "datatype": "string"}]}, "qualifiers-order": ["P4390", "P1810"], "id": "Q72638$7C986135-C8B7-4E77-B4BE-5E0B8A4CE82D", "rank": "normal"}], "P1552": [{"mainsnak": {"snaktype": "value", "property": "P1552", "hash": "b135150c44a47324166e0712bfa44f0dad593f23", "datavalue": {"value": {"entity-type": "item", "numeric-id": 1053008, "id": "Q1053008"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q72638$c51063fb-4566-b641-32c6-a57c5ea8dce0", "rank": "normal"}]}}}, "success": 1}
