+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q7887142.json
@@ -1,0 +1,1 @@
+{"entities": {"Q7887142": {"pageid": 7824318, "ns": 0, "title": "Q7887142", "lastrevid": 2049060194, "modified": "2024-01-08T19:53:32Z", "type": "item", "id": "Q7887142", "labels": {"en": {"language": "en", "value": "unit of analysis"}}, "descriptions": {"en": {"language": "en", "value": "major entity that is being analyzed in a study"}}, "aliases": {}, "claims": {"P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "7d0af84aa58c62f2e0dfa5ee073023716ec07c81", "datavalue": {"value": {"entity-type": "item", "numeric-id": 4330518, "id": "Q4330518"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q7887142$6178745a-4fe2-5e1e-9df7-04f3a48706b7", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P279", "hash": "39b3f3f55b83550fc005b051c8a8bafae58c7039", "datavalue": {"value": {"entity-type": "item", "numeric-id": 2198779, "id": "Q2198779"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q7887142$62aa181e-499f-ee17-fcb9-0cb917792f47", "rank": "normal"}], "P6366": [{"mainsnak": {"snaktype": "value", "property": "P6366", "hash": "025cddf9a4228a5b4b14b604993a1b0592acdcf4", "datavalue": {"value": "54797493", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q7887142$E416AB21-1CDE-4B1D-A238-3816660D37FE", "rank": "normal"}], "P646": [{"mainsnak": {"snaktype": "value", "property": "P646", "hash": "eae58756b53922006bfc0073a542d60d7b852b0f", "datavalue": {"value": "/m/0blk7l", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q7887142$EB63AB92-582F-4F8B-BC90-4146038C050E", "rank": "normal"}]}}}, "success": 1}
