+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q902805.json
@@ -1,0 +1,1 @@
+{"entities": {"Q902805": {"pageid": 853571, "ns": 0, "title": "Q902805", "lastrevid": 2320776767, "modified": "2025-03-06T13:03:33Z", "type": "item", "id": "Q902805", "labels": {"en": {"language": "en", "value": "primary nutritional group"}}, "descriptions": {"en": {"language": "en", "value": "group of organisms, divided in relation to the nutrition mode according to the sources of energy and carbon, needed for living, growth and reproduction"}}, "aliases": {"en": [{"language": "en", "value": "trophism"}]}, "claims": {"P646": [{"mainsnak": {"snaktype": "value", "property": "P646", "hash": "2bde032eeb9eb1124a3d622e0050152d24322a01", "datavalue": {"value": "/m/0p11j", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q902805$8098B2CE-2921-4769-8030-EBB0A7C2165C", "rank": "normal", "references": [{"hash": "2b00cb481cddcac7623114367489b5c194901c4a", "snaks": {"P248": [{"snaktype": "value", "property": "P248", "hash": "a94b740202b097dd33355e0e6c00e54b9395e5e0", "datavalue": {"value": {"entity-type": "item", "numeric-id": 15241312, "id": "Q15241312"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}], "P577": [{"snaktype": "value", "property": "P577", "hash": "fde79ecb015112d2f29229ccc1ec514ed3e71fa2", "datavalue": {"value": {"time": "+2013-10-28T00:00:00Z", "timezone": 0, "before": 0, "after": 0, "precision": 11, "calendarmodel": "http://www.wikidata.org/entity/Q1985727"}, "type": "time"}, "datatype": "time"}]}, "snaks-order": ["P248", "P577"]}]}], "P1269": [{"mainsnak": {"snaktype": "value", "property": "P1269", "hash": "0a3c7c75226ab93a1b2a7992996b824557575d44", "datavalue": {"value": {"entity-type": "item", "numeric-id": 42922451, "id": "Q42922451"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q902805$17a9fe05-4090-4d2f-d2c4-19e4ed121602", "rank": "normal"}], "P910": [{"mainsnak": {"snaktype": "value", "property": "P910", "hash": "1a5748fa3588bf22cec0debdc4b8675f4e930378", "datavalue": {"value": {"entity-type": "item", "numeric-id": 32127149, "id": "Q32127149"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q902805$B7593F75-3C8A-440A-856A-229C65FAFDB4", "rank": "normal", "references": [{"hash": "d4bd87b862b12d99d26e86472d44f26858dee639", "snaks": {"P143": [{"snaktype": "value", "property": "P143", "hash": "f30cbd35620c4ea6d0633aaf0210a8916130469b", "datavalue": {"value": {"entity-type": "item", "numeric-id": 8447, "id": "Q8447"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "snaks-order": ["P143"]}]}], "P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "cc1eb50276462e05cca1d1bb823266f6cc24073c", "datavalue": {"value": {"entity-type": "item", "numeric-id": 7239, "id": "Q7239"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P1013": [{"snaktype": "value", "property": "P1013", "hash": "2088038a58dcdb2bae6d4641af24fa834f79985a", "datavalue": {"value": {"entity-type": "item", "numeric-id": 159462, "id": "Q159462"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, {"snaktype": "value", "property": "P1013", "hash": "d66b317265e81665f219e2cd41b358fd5206b682", "datavalue": {"value": {"entity-type": "item", "numeric-id": 2138622, "id": "Q2138622"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P1013"], "id": "Q902805$3eefa826-47a0-a49c-abd4-826a457defd9", "rank": "deprecated"}, {"mainsnak": {"snaktype": "value", "property": "P279", "hash": "c43b627d5fcf50f57bf2c6106ab142904b2639c4", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24017414, "id": "Q24017414"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q902805$423e17bb-4214-03f4-f91c-05cd911a2206", "rank": "normal"}], "P6366": [{"mainsnak": {"snaktype": "value", "property": "P6366", "hash": "abb91f870dd9714da753081872dab7357e2d6b8e", "datavalue": {"value": "117057483", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q902805$7F02032E-BCFC-45CF-96F4-FEF3268C457B", "rank": "normal"}], "P8408": [{"mainsnak": {"snaktype": "value", "property": "P8408", "hash": "3ad7eb09f88961937d1753384d19f7d6cb397322", "datavalue": {"value": "Photolithotroph", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q902805$BC478E68-4111-43AA-82E8-F9AD262EC961", "rank": "normal", "references": [{"hash": "9a681f9dd95c90224547c404e11295f4f7dcf54e", "snaks": {"P248": [{"snaktype": "value", "property": "P248", "hash": "9d5780dddffa8746637a9929a936ab6b0f601e24", "datavalue": {"value": {"entity-type": "item", "numeric-id": 64139102, "id": "Q64139102"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}], "P813": [{"snaktype": "value", "property": "P813", "hash": "622a5a27fa5b25e7e7984974e9db494cf8460990", "datavalue": {"value": {"time": "+2020-07-09T00:00:00Z", "timezone": 0, "before": 0, "after": 0, "precision": 11, "calendarmodel": "http://www.wikidata.org/entity/Q1985727"}, "type": "time"}, "datatype": "time"}]}, "snaks-order": ["P248", "P813"]}]}], "P8225": [{"mainsnak": {"snaktype": "value", "property": "P8225", "hash": "c42b95c1071479a6438d6fc54b529596a9c24bd1", "datavalue": {"value": {"entity-type": "item", "numeric-id": 7239, "id": "Q7239"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P1013": [{"snaktype": "value", "property": "P1013", "hash": "2088038a58dcdb2bae6d4641af24fa834f79985a", "datavalue": {"value": {"entity-type": "item", "numeric-id": 159462, "id": "Q159462"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, {"snaktype": "value", "property": "P1013", "hash": "d66b317265e81665f219e2cd41b358fd5206b682", "datavalue": {"value": {"entity-type": "item", "numeric-id": 2138622, "id": "Q2138622"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P1013"], "id": "Q902805$edfdbf10-4c8a-ff47-f326-d71e0d335b7e", "rank": "normal"}]}}}, "success": 1}
