+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q931483.json
@@ -1,0 +1,1 @@
+{"entities": {"Q931483": {"pageid": 881802, "ns": 0, "title": "Q931483", "lastrevid": 2296427575, "modified": "2025-01-09T11:07:53Z", "type": "item", "id": "Q931483", "labels": {"en": {"language": "en", "value": "legal fiction"}}, "descriptions": {"en": {"language": "en", "value": "fact assumed or created by courts which is then used in order to apply a legal rule"}}, "aliases": {"en": [{"language": "en", "value": "fiction"}]}, "claims": {"P910": [{"mainsnak": {"snaktype": "value", "property": "P910", "hash": "f6fb4b5fa4c33fbf86c9edd02215b3c73d559dff", "datavalue": {"value": {"entity-type": "item", "numeric-id": 8586652, "id": "Q8586652"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q931483$93F02E01-52FF-4C0F-B42F-FE2A874F5D47", "rank": "normal"}], "P646": [{"mainsnak": {"snaktype": "value", "property": "P646", "hash": "4caa986e5f2c6e62254d40a959c0457577179ed0", "datavalue": {"value": "/m/01m4r0", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q931483$4D3D579E-092B-4DF7-AECA-B236D9573A7B", "rank": "normal", "references": [{"hash": "2b00cb481cddcac7623114367489b5c194901c4a", "snaks": {"P248": [{"snaktype": "value", "property": "P248", "hash": "a94b740202b097dd33355e0e6c00e54b9395e5e0", "datavalue": {"value": {"entity-type": "item", "numeric-id": 15241312, "id": "Q15241312"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}], "P577": [{"snaktype": "value", "property": "P577", "hash": "fde79ecb015112d2f29229ccc1ec514ed3e71fa2", "datavalue": {"value": {"time": "+2013-10-28T00:00:00Z", "timezone": 0, "before": 0, "after": 0, "precision": 11, "calendarmodel": "http://www.wikidata.org/entity/Q1985727"}, "type": "time"}, "datatype": "time"}]}, "snaks-order": ["P248", "P577"]}]}], "P1343": [{"mainsnak": {"snaktype": "value", "property": "P1343", "hash": "eda63bed5c3d7a3460033092338ab321a2374c7f", "datavalue": {"value": {"entity-type": "item", "numeric-id": 2041543, "id": "Q2041543"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P805": [{"snaktype": "value", "property": "P805", "hash": "50c512b453e0e2894c412bbf8081663908edb87e", "datavalue": {"value": {"entity-type": "item", "numeric-id": 23857837, "id": "Q23857837"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P805"], "id": "Q931483$4B774A02-0569-47E4-8011-09D1E5033385", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P1343", "hash": "6f3a47f16e2e5263b8d3d7b75ade9ed1a81c39f2", "datavalue": {"value": {"entity-type": "item", "numeric-id": 108732786, "id": "Q108732786"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P1810": [{"snaktype": "value", "property": "P1810", "hash": "d71852a06e14f338b82b6fddeadd05119989fbb6", "datavalue": {"value": "legal fiction", "type": "string"}, "datatype": "string"}]}, "qualifiers-order": ["P1810"], "id": "Q931483$d1f2da77-465f-406d-b3ca-78486b78fc36", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P1343", "hash": "1e87e0f7d2db06ab0ef5ccabd7c6275641f36553", "datavalue": {"value": {"entity-type": "item", "numeric-id": 867541, "id": "Q867541"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "qualifiers": {"P805": [{"snaktype": "value", "property": "P805", "hash": "dde27ca0bb6da822fb7678ee089d65666f294101", "datavalue": {"value": {"entity-type": "item", "numeric-id": 84688475, "id": "Q84688475"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "qualifiers-order": ["P805"], "id": "Q931483$830AA1D3-B11D-47A3-BDCF-2E3DFDEFEE48", "rank": "normal"}], "P3417": [{"mainsnak": {"snaktype": "value", "property": "P3417", "hash": "e086a47dad08e47238339a139cbc587245d82826", "datavalue": {"value": "Legal-Fiction", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q931483$D33A5337-A20E-40CA-9542-6666A57B5AFD", "rank": "normal", "references": [{"hash": "3b0a5bb3c1f955edce73740124f7d935698092ad", "snaks": {"P248": [{"snaktype": "value", "property": "P248", "hash": "3ac9682e789a3a3791d4fd088b265ea03abef101", "datavalue": {"value": {"entity-type": "item", "numeric-id": 51711, "id": "Q51711"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}]}, "snaks-order": ["P248"]}]}], "P3827": [{"mainsnak": {"snaktype": "value", "property": "P3827", "hash": "52a1e3abec938df41d649dad936307b02e554023", "datavalue": {"value": "legal-fictions", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q931483$0A545BB2-425F-466D-904D-C7FC7C066558", "rank": "normal"}], "P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "ce18f6222d4420135dc2b5186efc3fa9019a8ab7", "datavalue": {"value": {"entity-type": "item", "numeric-id": 2135465, "id": "Q2135465"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q931483$b6bb2838-4f63-9ea4-9105-38d86c49df53", "rank": "normal"}], "P1417": [{"mainsnak": {"snaktype": "value", "property": "P1417", "hash": "ff22ee2771c29fdf3b6a44203464d48cbe827fde", "datavalue": {"value": "topic/legal-fiction", "type": "string"}, "datatype": "external-id"}, "type": "statement", "qualifiers": {"P1810": [{"snaktype": "value", "property": "P1810", "hash": "d71852a06e14f338b82b6fddeadd05119989fbb6", "datavalue": {"value": "legal fiction", "type": "string"}, "datatype": "string"}]}, "qualifiers-order": ["P1810"], "id": "Q931483$D4B4B440-5144-4BF7-8757-C598F9FC7EEE", "rank": "normal"}], "P6366": [{"mainsnak": {"snaktype": "value", "property": "P6366", "hash": "e5880502b6aeb2507df7b58e3c93753c768e81f7", "datavalue": {"value": "2777992059", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q931483$47909347-BE7E-4FF0-BFC5-F133EDE6E70D", "rank": "normal"}], "P10283": [{"mainsnak": {"snaktype": "value", "property": "P10283", "hash": "c7456bb5ccc26e93efed8e73c9870806f22b7795", "datavalue": {"value": "C2777992059", "type": "string"}, "datatype": "external-id"}, "type": "statement", "id": "Q931483$58A80745-E496-4E11-A053-EC396648E8E3", "rank": "normal", "references": [{"hash": "c13ba4d7902ead2693f79396bc25bd35d4d14030", "snaks": {"P248": [{"snaktype": "value", "property": "P248", "hash": "4a4f26a5361b5707266e48e425bf2be2f99fd2ab", "datavalue": {"value": {"entity-type": "item", "numeric-id": 107507571, "id": "Q107507571"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}], "P813": [{"snaktype": "value", "property": "P813", "hash": "435834d08182bb9f3dbe974ba9840af0f12899cc", "datavalue": {"value": {"time": "+2022-01-26T00:00:00Z", "timezone": 0, "before": 0, "after": 0, "precision": 11, "calendarmodel": "http://www.wikidata.org/entity/Q1985727"}, "type": "time"}, "datatype": "time"}], "P854": [{"snaktype": "value", "property": "P854", "hash": "a4a4bf53f22268815c51ec10fed608da703c9c7f", "datavalue": {"value": "https://docs.openalex.org/download-snapshot/snapshot-data-format", "type": "string"}, "datatype": "url"}]}, "snaks-order": ["P248", "P813", "P854"]}]}]}}}, "success": 1}
