+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q98119401.json
@@ -1,0 +1,1 @@
+{"entities": {"Q98119401": {"pageid": 96402958, "ns": 0, "title": "Q98119401", "lastrevid": 2314183870, "modified": "2025-02-20T09:10:18Z", "type": "item", "id": "Q98119401", "labels": {"en": {"language": "en", "value": "group or class of physical objects"}}, "descriptions": {}, "aliases": {}, "claims": {"P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "0bf0845308c3278652bae47f49ae60c2a096d622", "datavalue": {"value": {"entity-type": "item", "numeric-id": 99527517, "id": "Q99527517"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q98119401$8649ce49-4fc7-f098-6bec-84c468ef8042", "rank": "normal"}, {"mainsnak": {"snaktype": "value", "property": "P279", "hash": "7f7dd87eb24b9d3e17980034bd7511358931d411", "datavalue": {"value": {"entity-type": "item", "numeric-id": 488383, "id": "Q488383"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q98119401$581afc95-49d5-c962-527a-5aae6f4a8c3f", "rank": "normal"}], "P31": [{"mainsnak": {"snaktype": "value", "property": "P31", "hash": "a7e32719e3a70a25fce867c2ad70c306d3521fc3", "datavalue": {"value": {"entity-type": "item", "numeric-id": 24017414, "id": "Q24017414"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q98119401$CE4CF50E-C8C4-4D17-85AC-DA9549C4783B", "rank": "normal"}], "P8225": [{"mainsnak": {"snaktype": "value", "property": "P8225", "hash": "0cbb8274a4bd6ec7b30e978a0f7086b693051664", "datavalue": {"value": {"entity-type": "item", "numeric-id": 223557, "id": "Q223557"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q98119401$efaec2ed-4c6a-886c-5bc0-0527dbba7b8b", "rank": "normal"}]}}}, "success": 1}
