+++ tests/fixtures/wikidata/api_php_action_wbgetentities_format_json_ids_q99527517.json
@@ -1,0 +1,1 @@
+{"entities": {"Q99527517": {"pageid": 97445404, "ns": 0, "title": "Q99527517", "lastrevid": 2306830602, "modified": "2025-02-04T13:50:54Z", "type": "item", "id": "Q99527517", "labels": {"en": {"language": "en", "value": "collective entity"}}, "descriptions": {"en": {"language": "en", "value": "entity that is a composite of other entities"}}, "aliases": {"en": [{"language": "en", "value": "composite entity"}, {"language": "en", "value": "aggregate entity"}, {"language": "en", "value": "collection entity"}, {"language": "en", "value": "group or class"}, {"language": "en", "value": "class or group"}]}, "claims": {"P31": [{"mainsnak": {"snaktype": "value", "property": "P31", "hash": "9d31d99ce62f7f1424d6d2be409f25a62397445f", "datavalue": {"value": {"entity-type": "item", "numeric-id": 23958852, "id": "Q23958852"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q99527517$5029db41-4229-75ed-0dc6-28e6f6ba9108", "rank": "normal"}], "P279": [{"mainsnak": {"snaktype": "value", "property": "P279", "hash": "07a161639702c3cda624d57463695bfc6c5439e3", "datavalue": {"value": {"entity-type": "item", "numeric-id": 35120, "id": "Q35120"}, "type": "wikibase-entityid"}, "datatype": "wikibase-item"}, "type": "statement", "id": "Q99527517$501e706c-42fc-1b81-dcdd-6beb07e0f9e6", "rank": "normal"}]}}}, "success": 1}
